refactor: dispatch typed storyline events through family handlers
