+++ scripts/fixtures/cleanup-inputs.json
@@ -1,0 +1,8 @@
+{
+  "checkpoint-all-formats.yaml": "80d2d66ebea1d1ee78d661104dd67a82c7566e6db6200f4b79c736de344c5c8e",
+  "cleanup-typed-handlers.yaml": "ddf9f9002a1f383219bccd07aea3e69c5f5cbaab51800f73dc306a1f9abab8be",
+  "full-coverage-apt.yaml": "75e30a3935661665885c9244da0fbee0bdeda899c596410da1cc2da4920890be",
+  "minimal.yaml": "49f2abc4d370291124ccd1a73048e3bb7a3662c84541d5a6cae21e831618605a",
+  "smb-linux-matrix.yaml": "6319136bee424c23a2739c102e847e09a9136e588c0119c76e95a8b0722848b8",
+  "smb-resource-calibration.yaml": "5cf690d94c5b34d72db0e71f3e796d16b46f537d461040347c54085b053b3d2a"
+}
