+++ src/evidenceforge/generation/engine/storyline.py
@@ -40,7 +40,6 @@ import re
 import shlex
 import string
 from collections.abc import Iterable, Iterator, Sequence
-from dataclasses import replace
 from datetime import UTC, datetime, timedelta
 from types import SimpleNamespace
 from typing import Any
@@ -51,15 +50,12 @@ from evidenceforge.events.network import SignaturePredicate
 from evidenceforge.generation.actions import (
     IdsAlertActionBundle,
     IdsAlertRequest,
-    PortScanActionBundle,
     PortScanRequest,
     ScpReceiverFileActionBundle,
     ScpReceiverFileRequest,
     StagedArchiveSmbReadActionBundle,
     StagedArchiveSmbReadRequest,
-    WebScanActionBundle,
     WebScanRequest,
-    dhcp_renewal_interval_seconds,
     dns_transport_close_headroom_seconds,
     network_transport_open_positive_headroom_seconds,
 )
@@ -80,7 +76,6 @@ from evidenceforge.generation.activity.http_content import (
     response_size_for_status,
 )
 from evidenceforge.generation.activity.network import _is_private_ip
-from evidenceforge.generation.activity.network_params import activity_dns_resolver_ips
 from evidenceforge.generation.intent_ledger import IntentSection
 from evidenceforge.generation.storage_world import CompiledStorageFile
 from evidenceforge.generation.world_model import (
@@ -93,7 +88,6 @@ from evidenceforge.models.exceptions import StateError
 from evidenceforge.models.ids import IdsAlertAttachmentSpec
 from evidenceforge.models.scenario import (
     MAX_HTTP_RESPONSE_BODY_LEN,
-    BeaconHttpSequenceEntry,
     ConnectionEventSpec,
     EventSpacingConfig,
     SmbClientLocation,
@@ -4450,3580 +4444,26 @@ class StorylineMixin:
                     return uid
             return "(filtered by sensor placement)"
 
-        if spec.type == "logon":
-            session_end_plan = self._session_end_plan_for_current_start()
-            if (
-                spec.logon_type == 10
-                and _get_os_category(system.os) == "windows"
-                and spec.source_ip not in {"-", system.ip}
-            ):
-                session_end_plan = self._authored_rdp_session_end_plan()
-            if spec.logon_type == 9:
-                caller, caller_logon_id = self._storyline_new_credentials_caller(
-                    actor,
-                    system,
-                    time,
-                )
-                outbound_domain = self.activity_generator._explicit_credentials_target_domain(
-                    actor.username,
-                    system.hostname,
-                    system,
-                )
-                logon_id = self.activity_generator._emit_new_credentials_logon(
-                    user=caller,
-                    system=system,
-                    time=time,
-                    caller_logon_id=caller_logon_id,
-                    outbound_username=actor.username,
-                    outbound_domain=outbound_domain,
-                    lifecycle_group_id=(
-                        f"storyline:{getattr(self, '_current_storyline_spec_id', 'logon')}"
-                    ),
-                )
-                source_ip = "-"
-            else:
-                source_ip = (
-                    spec.source_ip
-                    or self.public_identity_registry.bind(
-                        "external_logon",
-                        f"storyline-logon:{actor.username}:{system.hostname}:{time.isoformat()}",
-                    ).ip
-                )
-                logon_id = self.activity_generator.generate_logon(
-                    user=actor,
-                    system=system,
-                    time=time,
-                    logon_type=spec.logon_type,
-                    source_ip=source_ip,
-                    session_end_plan=session_end_plan,
-                )
-            # Protect storyline-created sessions from baseline logoff
-            session = self.state_manager.get_session(logon_id)
-            if session:
-                session.storyline_protected = True
-                if getattr(session, "session_kind", "") in {"rdp", "ssh"}:
-                    self._record_storyline_session_ready(
-                        system=system,
-                        actor=actor,
-                        session=session,
-                        rng=rng,
-                    )
-            malicious_event["logon_id"] = logon_id
-            malicious_event["source_ip"] = source_ip
-            self._record_storyline_logon(actor, system, logon_id, source_ip=source_ip)
+        from .typed_handlers import TypedEventContext, execute_typed_handler
 
-        elif spec.type == "failed_logon":
-            source_ip = (
-                spec.source_ip
-                or self.public_identity_registry.bind(
-                    "failed_logon",
-                    f"storyline-failed-logon:{actor.username}:{system.hostname}:{time.isoformat()}",
-                ).ip
-            )
-            dc = next(
-                (s for s in self.scenario.environment.systems if s.type == "domain_controller"),
-                None,
-            )
-            self.activity_generator.generate_failed_logon(
-                user=actor,
-                system=system,
-                time=time,
-                logon_type=spec.logon_type,
-                source_ip=source_ip,
-                target_username=getattr(spec, "target_username", None),
-                dc_system=dc,
-            )
-            malicious_event["source_ip"] = source_ip
-
-        elif spec.type == "logoff":
-            paired_close = self._session_end_plan_for_current_logoff()
-            if paired_close is not None:
-                logon_id, session_end_plan = paired_close
-                target_session = self.state_manager.get_session(logon_id)
-                if target_session is None or target_session.system != system.hostname:
-                    raise StateError(
-                        f"Explicit storyline logoff {session_end_plan.storyline_event_id} "
-                        f"lost its paired session {logon_id} on {system.hostname}"
-                    )
-                self.activity_generator.generate_logoff(
-                    actor,
-                    system,
-                    session_end_plan.canonical_end,
-                    logon_id,
-                    from_storyline=True,
-                    session_end_plan=session_end_plan,
-                )
-            else:
-                logon_id = self._last_storyline_logon_for_actor_system(
-                    actor,
-                    system,
-                    at_time=time,
-                )
-                if logon_id is not None:
-                    self.activity_generator.generate_logoff(
-                        actor,
-                        system,
-                        time,
-                        logon_id,
-                        from_storyline=True,
-                    )
-
-        elif spec.type == "email_message":
-            result = self.activity_generator.generate_email_message(
-                spec=spec,
+        return execute_typed_handler(
+            self,
+            spec,
+            TypedEventContext(
                 actor=actor,
                 system=system,
                 time=time,
                 activity=activity,
-                storyline_id=getattr(dispatcher, "storyline_cluster_id", "") or "",
-            )
-            malicious_event.update(
-                {
-                    "artifact_id": result.artifact_id,
-                    "message_id": result.message_id,
-                    "sender": result.sender,
-                    "recipients": result.recipients,
-                    "subject": result.subject,
-                    "outcome": result.outcome,
-                    "artifact_path": result.artifact_path,
-                    "smtp_uids": result.smtp_uids,
-                    "route": result.route,
-                }
-            )
-
-        elif spec.type == "email_read":
-            result = self.activity_generator.generate_email_read(
-                spec=spec,
-                actor=actor,
-                system=system,
-                time=time,
-                activity=activity,
-                storyline_id=getattr(dispatcher, "storyline_cluster_id", "") or "",
-            )
-            malicious_event.update(result)
-
-        elif spec.type == "process":
-            os_category = _get_os_category(system.os)
-            if hasattr(self, "world_planner"):
-                # Built-in/service accounts (SYSTEM, LOCAL SERVICE, etc.) run
-                # locally — don't fabricate remote logon evidence for them.
-                from evidenceforge.validation.schema import BUILTIN_ACCOUNTS
-
-                service_accounts = set(self.scenario.environment.service_accounts)
-                is_local_account = (
-                    actor.username in BUILTIN_ACCOUNTS or actor.username in service_accounts
-                )
-                is_interactive_linux_root = os_category == "linux" and actor.username == "root"
-                if is_local_account and not is_interactive_linux_root:
-                    linux_daemon_users = {"apache", "www-data", "nginx", "httpd", "tomcat"}
-                    if os_category == "linux" and actor.username.lower() in linux_daemon_users:
-                        logon_id = ""
-                    else:
-                        # Use existing system session or create a service logon.
-                        sessions = self.state_manager.get_sessions_for_user_at(
-                            actor.username,
-                            time,
-                        )
-                        target_session = max(
-                            (s for s in sessions if s.system == system.hostname),
-                            key=lambda session: session.start_time,
-                            default=None,
-                        )
-                        if target_session:
-                            logon_id = target_session.logon_id
-                        else:
-                            logon_time = time - timedelta(seconds=rng.uniform(0.5, 2.0))
-                            logon_id = self.activity_generator.generate_service_logon(
-                                system=system,
-                                time=logon_time,
-                                service_account=actor.username,
-                            )
-                else:
-                    logon_id = self._last_storyline_logon_for_actor_system(
-                        actor,
-                        system,
-                        at_time=time,
-                    )
-                    if logon_id is None:
-                        required_until = self._next_storyline_logoff_time_for_actor_system(
-                            actor,
-                            system,
-                            time,
-                        )
-                        session_kind = self._storyline_non_session_kind(actor, system, rng)
-                        target_session = self.world_planner.ensure_user_session(
-                            actor,
-                            system,
-                            time,
-                            rng,
-                            session_kind=session_kind,
-                            storyline_protected=True,
-                            required_until=required_until,
-                        )
-                        logon_id = target_session.logon_id
-                        self._record_storyline_logon(actor, system, logon_id)
-            else:
-                sessions = self.state_manager.get_sessions_for_user(actor.username)
-                target_session = max(
-                    (s for s in sessions if s.system == system.hostname),
-                    key=lambda session: session.start_time,
-                    default=None,
-                )
-                if not target_session:
-                    logon_time = time - timedelta(seconds=rng.uniform(0.5, 2.0))
-                    logon_id = self.activity_generator.generate_logon(
-                        actor, system, logon_time, logon_type=3
-                    )
-                    self._record_storyline_logon(actor, system, logon_id)
-                else:
-                    logon_id = target_session.logon_id
-
-            process_actor = self._linux_native_service_user_for_storyline_actor(
-                actor,
-                system,
-                time,
-            )
-            process_actor = self._storyline_local_process_actor_for_logon(
-                process_actor,
-                system,
-                logon_id,
-            )
-            process_name = _normalize_storyline_process_image(
-                spec.process_name,
-                os_category,
-                username=process_actor.username,
-            )
-            command_line = spec.command_line or process_name
-            shell_key = (system.hostname, process_actor.username)
-
-            if os_category == "linux":
-                if not hasattr(self, "_storyline_shell_available_at"):
-                    self._storyline_shell_available_at: dict[tuple[str, str], datetime] = {}
-                available_times = [
-                    ts
-                    for key in {shell_key, (system.hostname, actor.username)}
-                    if (ts := self._storyline_shell_available_at.get(key)) is not None
-                ]
-                available_at = max(available_times) if available_times else None
-                if available_at is not None and time < available_at:
-                    time = available_at + timedelta(seconds=rng.uniform(0.3, 2.0))
-
-            if "<base64_encoded_command>" in command_line:
-                command_line = command_line.replace(
-                    "<base64_encoded_command>",
-                    self._generate_encoded_powershell(
-                        _stable_seed(f"storyline_ps_{time.isoformat()}_{actor.username}")
-                    ),
-                )
-
-            process_command_line = command_line
-            if os_category == "linux":
-                from evidenceforge.generation.activity.generator import (
-                    _linux_command_process_from_shell,
-                )
-
-                inferred_process = _linux_command_process_from_shell(
-                    command_line,
-                    username=process_actor.username,
-                )
-                if inferred_process is not None:
-                    inferred_image, inferred_command_line = inferred_process
-                    if inferred_image.rsplit("/", 1)[-1] == process_name.rsplit("/", 1)[-1]:
-                        process_command_line = inferred_command_line
-                shell_command_line = _linux_shell_process_command_line(
-                    process_name,
-                    process_command_line,
-                )
-                if shell_command_line is not None:
-                    process_command_line = shell_command_line
-
-            output_file = self._extract_output_file(command_line, os_category)
-            process_logon_id = logon_id
-            service_lifecycle_group_id = ""
-            service_process_identity = self._storyline_service_process_identity(
-                system=system,
-                time=time,
-                process_name=process_name,
+                explicit_types=explicit_types,
                 future_specs=future_specs,
-            )
-            parent_ref = getattr(spec, "parent_ref", None)
-            if not isinstance(parent_ref, str) or not parent_ref:
-                parent_ref = None
-            explicit_parent = self._storyline_process_ref_for_parent(
-                actor=process_actor,
-                system=system,
-                parent_ref=parent_ref,
-            )
-            if parent_ref is not None and explicit_parent is None:
-                malicious_event["process_name"] = process_name
-                malicious_event["command_line"] = command_line
-                malicious_event["skipped_reason"] = "no_live_parent_ref"
-                return malicious_event
-            if explicit_parent is not None:
-                parent_pid, _parent_image = explicit_parent
-            elif service_process_identity is not None:
-                process_actor, _service_name, service_lifecycle_group_id = service_process_identity
-                process_logon_id = {
-                    "SYSTEM": "0x3e7",
-                    "LOCAL SERVICE": "0x3e5",
-                    "NETWORK SERVICE": "0x3e4",
-                }[process_actor.username]
-                parent_pid = self.activity_generator._get_system_pid(
-                    system.hostname,
-                    "services",
-                    0x1F4,
-                )
-            else:
-                service_context = self._storyline_service_context_for_process(
-                    actor=process_actor,
-                    system=system,
-                    time=time,
-                    process_name=process_name,
-                )
-                if service_context is not None:
-                    (
-                        process_actor,
-                        process_logon_id,
-                        parent_pid,
-                        service_lifecycle_group_id,
-                    ) = service_context
-                else:
-                    parent_pid = self.activity_generator._resolve_parent(
-                        system,
-                        process_actor,
-                        time,
-                        process_logon_id,
-                        process_name,
-                        process_command_line,
-                    )
-            if os_category == "linux":
-                self._emit_linux_storyline_shell_friction(
-                    actor=process_actor,
-                    system=system,
-                    time=time,
-                    process_name=process_name,
-                    command_line=command_line,
-                    output_file=output_file,
-                    rng=rng,
-                )
-                reserved_start_time = (
-                    self.activity_generator.reserve_linux_foreground_process_start(
-                        system=system,
-                        username=process_actor.username,
-                        logon_id=process_logon_id,
-                        parent_pid=parent_pid,
-                        requested_time=time,
-                        process_name=process_name,
-                        command_line=process_command_line,
-                        authoritative_time=True,
-                    )
-                )
-                if isinstance(reserved_start_time, datetime):
-                    time = reserved_start_time
-                prepared_shell_command = self.activity_generator._prepare_bash_history_command(
-                    system,
-                    command_line,
-                )
-                self.activity_generator._emit_bash_command_event(
-                    process_actor,
-                    system,
-                    time,
-                    prepared_shell_command,
-                )
-                malicious_event["time"] = time
-            exe_name = process_name.rsplit("\\", 1)[-1].rsplit("/", 1)[-1].lower()
-            service_backed_process = service_process_identity is not None or (
-                "service_installed" in explicit_types
-                and exe_name in {"psexesvc.exe", "healthmonitorsvc.exe"}
-            )
-            if service_backed_process and not service_lifecycle_group_id:
-                matching_service_spec = next(
-                    (
-                        candidate
-                        for candidate in future_specs
-                        if getattr(candidate, "type", "") == "service_installed"
-                        and self._normalize_storyline_service_file_name(
-                            str(getattr(candidate, "service_file_name", "") or "")
-                        )
-                        .rsplit("\\", 1)[-1]
-                        .casefold()
-                        == exe_name
-                    ),
-                    None,
-                )
-                if matching_service_spec is not None:
-                    service_name = str(
-                        getattr(matching_service_spec, "service_name", "") or exe_name
-                    )
-                else:
-                    service_name = exe_name.removesuffix(".exe")
-                service_lifecycle_group_id = self._storyline_remote_service_lifecycle_id(
-                    system,
-                    service_name,
-                )
-            if not service_lifecycle_group_id and exe_name in {
-                "psexesvc.exe",
-                "healthmonitorsvc.exe",
-            }:
-                recent_service = getattr(self, "_last_storyline_service_by_system", {}).get(
-                    system.hostname,
-                    {},
-                )
-                installed_image = self._normalize_storyline_service_file_name(
-                    str(recent_service.get("service_file_name") or "")
-                )
-                installed_at = recent_service.get("installed_at")
-                if (
-                    installed_image.rsplit("\\", 1)[-1].casefold() == exe_name
-                    and isinstance(installed_at, datetime)
-                    and installed_at <= time
-                    and time - installed_at
-                    <= (
-                        timedelta(minutes=2)
-                        if exe_name == "psexesvc.exe"
-                        else timedelta(minutes=30)
-                    )
-                ):
-                    service_lifecycle_group_id = str(recent_service.get("lifecycle_group_id") or "")
-            pid = self.activity_generator.generate_process(
-                user=process_actor,
-                system=system,
-                time=time,
-                logon_id=process_logon_id,
-                process_name=process_name,
-                command_line=process_command_line,
-                parent_pid=parent_pid,
-                require_exact_parent=explicit_parent is not None,
-                ensure_file_event=not service_backed_process,
-                from_storyline=True,
-                suppress_command_file_effect=output_file is not None,
-                lifecycle_group_id=service_lifecycle_group_id,
-            )
-            running_process = self.state_manager.get_process(system.hostname, pid)
-            if running_process is not None:
-                time = max(ensure_utc(time), ensure_utc(running_process.start_time))
-                malicious_event["time"] = time
-            self.activity_generator._record_user_process(system, process_actor, pid, process_name)
-            self._record_last_storyline_process(system, pid, process_name, process_command_line)
-            process_ref = getattr(spec, "process_ref", None)
-            if not isinstance(process_ref, str) or not process_ref:
-                process_ref = None
-            if process_ref is not None:
-                self._record_storyline_process_ref(
-                    actor=process_actor,
-                    system=system,
-                    process_ref=process_ref,
-                    pid=pid,
-                    image=process_name,
-                )
-            malicious_event["process_name"] = process_name
-            malicious_event["command_line"] = command_line
-            malicious_event["pid"] = pid
-            archive_destination = self._extract_compress_archive_destination(command_line)
-            if archive_destination:
-                staging_source_ip = self._last_storyline_logon_source_for_actor_system(
-                    actor,
-                    system,
-                    at_time=time,
-                )
-                if staging_source_ip and staging_source_ip != system.ip:
-                    self._record_storyline_staged_archive(
-                        actor=process_actor,
-                        system=system,
-                        archive_path=archive_destination,
-                        source_ip=staging_source_ip,
-                        staged_at=time,
-                    )
-                    malicious_event["staged_archive"] = archive_destination
-            task_name = _extract_schtasks_option(command_line, "tn")
-            if task_name and "/create" in command_line.lower():
-                self._record_storyline_scheduled_task_command(system, task_name, command_line)
-            self._record_storyline_service_create_command(system, command_line)
-            self._record_storyline_account_create_command(system, command_line)
-
-            if output_file:
-                if os_category == "linux" and output_file.startswith("~/"):
-                    home = (
-                        "/root"
-                        if process_actor.username == "root"
-                        else f"/home/{process_actor.username}"
-                    )
-                    output_file = f"{home}/{output_file[2:]}"
-                file_time = time + timedelta(seconds=rng.uniform(0.5, 3.0))
-                from evidenceforge.events.base import OccurrenceBuilder
-                from evidenceforge.events.contexts import (
-                    AuthContext,
-                    FileContext,
-                    ProcessContext,
-                )
-
-                host_ctx = self.activity_generator._build_host_context(system)
-                running_proc = self.state_manager.get_process(system.hostname, pid)
-                self.dispatcher.dispatch_builder(
-                    OccurrenceBuilder(
-                        timestamp=file_time,
-                        event_type="file_create",
-                        src_host=host_ctx,
-                        auth=AuthContext(username=process_actor.username),
-                        process=ProcessContext(
-                            pid=pid,
-                            parent_pid=parent_pid,
-                            image=process_name,
-                            command_line=process_command_line,
-                            username=process_actor.username,
-                            logon_id=process_logon_id,
-                            start_time=running_proc.start_time
-                            if running_proc is not None
-                            else None,
-                        ),
-                        file=FileContext(path=output_file, action="create", pid=pid),
-                        storyline_origin=True,
-                    )
-                )
-                malicious_event["output_file"] = output_file
-
-            http_url = self._extract_http_url(command_line)
-            if http_url is not None:
-                parsed_target = self._parse_http_url_target(http_url)
-                if parsed_target is not None:
-                    from urllib.parse import urlparse
-
-                    from evidenceforge.events.contexts import HttpContext
-
-                    hostname, dst_port = parsed_target
-                    parsed_url = urlparse(http_url)
-                    uri = parsed_url.path or "/"
-                    if parsed_url.query:
-                        uri = f"{uri}?{parsed_url.query}"
-                    mime_type = normalize_mime_type_for_path(uri, "text/plain")
-                    response_body_len = (
-                        apply_transfer_size_variance(
-                            response_size_for_status(200, hostname, uri),
-                            status_code=200,
-                            host=hostname,
-                            uri=uri,
-                            content_type=mime_type,
-                            variant_key=f"{system.ip}:{process_name}:{pid}",
-                        )
-                        if is_stable_resource_path(uri)
-                        else response_size_for_mime(rng, mime_type)
-                    )
-                    preserve_url_dst_ip = False
-                    dst_ip = self._resolve_storyline_network_target(hostname)
-                    if dst_ip is None:
-                        authored_dst_ip = self._storyline_authored_ip_for_hostname(hostname)
-                        if authored_dst_ip is not None:
-                            dst_ip = authored_dst_ip
-                            preserve_url_dst_ip = True
-                    if dst_ip is None:
-                        dst_ip = self._resolve_scenario_network_host(
-                            hostname,
-                            src_host=system.hostname,
-                        )
-                    service = "ssl" if dst_port == 443 else "http"
-                    network_time = self._clamp_after_storyline_process_source_create(
-                        system=system,
-                        pid=pid,
-                        network_time=time + timedelta(milliseconds=rng.randint(250, 900)),
-                        rng=rng,
-                    )
-                    http_user_agent, http_user_agent_known_absent = (
-                        self._storyline_http_user_agent_metadata_for_command(
-                            system=system,
-                            process_image=process_name,
-                            command_line=command_line,
-                            rng=rng,
-                        )
-                    )
-                    self.activity_generator.generate_connection(
-                        src_ip=system.ip,
-                        dst_ip=dst_ip,
-                        time=network_time,
-                        dst_port=dst_port,
-                        proto="tcp",
-                        service=service,
-                        duration=rng.uniform(0.8, 6.0),
-                        orig_bytes=rng.randint(300, 1400),
-                        resp_bytes=response_body_len,
-                        conn_state="SF",
-                        emit_dns=not _is_private_ip(dst_ip),
-                        source_system=system,
-                        pid=pid,
-                        hostname=hostname,
-                        process_image=process_name,
-                        preserve_dst_ip=preserve_url_dst_ip,
-                        http=HttpContext(
-                            method="GET",
-                            host=hostname,
-                            uri=uri,
-                            version="1.1",
-                            user_agent=http_user_agent,
-                            user_agent_known_absent=http_user_agent_known_absent,
-                            request_body_len=0,
-                            response_body_len=response_body_len,
-                            status_code=200,
-                            status_msg="OK",
-                            resp_mime_types=[mime_type],
-                            tags=[],
-                        ),
-                    )
-                    malicious_event["network_url"] = http_url
-
-            remote_db_target = self._extract_database_client_target(command_line, os_category)
-            if remote_db_target is not None:
-                target_host, dst_port, service = remote_db_target
-                target_ip = self._resolve_storyline_network_target(target_host)
-                target_hostname = None if _IPV4_LITERAL_RE.fullmatch(target_host) else target_host
-                unresolved_single_label_fallback = False
-                if target_ip is None and target_hostname is not None:
-                    ad_domain = getattr(self, "_ad_domain", "")
-                    target_lower = target_hostname.rstrip(".").lower()
-                    unresolved_single_label = "." not in target_lower
-                    looks_internal = target_lower.endswith(".local") or (
-                        bool(ad_domain) and target_lower.endswith(f".{ad_domain.lower()}")
-                    )
-                    if unresolved_single_label:
-                        if self._is_local_database_instance_target(target_hostname):
-                            target_hostname = None
-                        else:
-                            target_ip = self._unresolved_database_target_ip(target_hostname)
-                            unresolved_single_label_fallback = target_ip is not None
-                            if ad_domain:
-                                target_hostname = f"{target_hostname}.{ad_domain}"
-                    elif not looks_internal:
-                        target_ip = self._resolve_scenario_network_host(
-                            target_hostname,
-                            src_host=system.hostname,
-                        )
-                if target_ip is not None:
-                    target_system = self._system_for_ip(target_ip)
-                    failed_private_attempt = unresolved_single_label_fallback or (
-                        target_system is None and _is_private_ip(target_ip)
-                    )
-                    firewall_ctx = None
-                    conn_state = "SF"
-                    duration = rng.uniform(0.6, 8.0)
-                    orig_bytes = rng.randint(180, 900)
-                    resp_bytes = rng.randint(800, 6000)
-                    rendered_service = service
-                    if failed_private_attempt:
-                        from evidenceforge.events.contexts import FirewallContext
-
-                        src_iface = self._resolve_firewall_interface(system.ip)
-                        dst_iface = self._resolve_firewall_interface(target_ip)
-                        firewall_ctx = FirewallContext(
-                            action="deny",
-                            msg_id=106023,
-                            connection_id=0,
-                            src_interface=src_iface,
-                            dst_interface=dst_iface,
-                            access_group=f"{src_iface}_access_in",
-                        )
-                        conn_state = self._get_firewall_deny_conn_state()
-                        duration = rng.uniform(0.02, 0.45)
-                        orig_bytes = 0
-                        resp_bytes = 0
-                        rendered_service = None
-                    connection_time = self._clamp_after_storyline_process_source_create(
-                        system=system,
-                        pid=pid,
-                        network_time=time + timedelta(milliseconds=rng.randint(250, 900)),
-                        rng=rng,
-                    )
-                    self.activity_generator.generate_connection(
-                        src_ip=system.ip,
-                        dst_ip=target_ip,
-                        time=connection_time,
-                        dst_port=dst_port,
-                        proto="tcp",
-                        service=rendered_service,
-                        duration=duration,
-                        orig_bytes=orig_bytes,
-                        resp_bytes=resp_bytes,
-                        conn_state=conn_state,
-                        emit_dns=target_hostname is not None,
-                        source_system=system,
-                        pid=pid,
-                        hostname=target_hostname,
-                        process_image=process_name,
-                        firewall=firewall_ctx,
-                    )
-                    malicious_event["network_target"] = target_host
-                    malicious_event["network_target_ip"] = target_ip
-                    malicious_event["network_target_port"] = dst_port
-
-            scp_destination = self._extract_scp_destination(command_line, os_category)
-            scp_target = scp_destination[0] if scp_destination is not None else None
-            if scp_target is not None:
-                dst_ip = self._resolve_storyline_network_target(scp_target)
-                if dst_ip:
-                    transfer_time = self._clamp_after_storyline_process_source_create(
-                        system=system,
-                        pid=pid,
-                        network_time=time + timedelta(milliseconds=rng.randint(250, 900)),
-                        rng=rng,
-                    )
-                    target_system = self._system_for_ip(dst_ip)
-                    modeled_linux_receiver = bool(
-                        target_system is not None and _get_os_category(target_system.os) == "linux"
-                    )
-                    source_port = (
-                        self.activity_generator.preview_ssh_source_port(
-                            system.ip,
-                            dst_ip,
-                            None,
-                            rng,
-                            _get_os_category(system.os),
-                            transfer_time,
-                        )
-                        if modeled_linux_receiver
-                        else self.activity_generator.reserve_ssh_source_port(
-                            system.ip,
-                            dst_ip,
-                            None,
-                            rng,
-                            _get_os_category(system.os),
-                            time=transfer_time,
-                        )
-                    )
-                    transfer_duration = (
-                        # Leave a real authentication window before exact SCP close.
-                        rng.uniform(12.0, 40.0)
-                        if modeled_linux_receiver
-                        else rng.uniform(2.0, 30.0)
-                    )
-                    source_path = self._extract_scp_source_path(command_line, os_category) or ""
-                    source_content = None
-                    runtime_manager = getattr(
-                        self.activity_generator,
-                        "_runtime_content_manager",
-                        None,
-                    )
-                    if runtime_manager is not None and source_path:
-                        source_record = runtime_manager.resolve_record(
-                            system.hostname,
-                            process_actor.username,
-                            source_path,
-                            "linux",
-                        )
-                        if source_record is not None:
-                            source_content = source_record.content
-                    if source_content is None and source_path:
-                        content_rng = random.Random(
-                            _stable_seed(
-                                "storyline_scp_source_content:"
-                                f"{system.hostname}:{source_path}:{transfer_time.isoformat()}"
-                            )
-                        )
-                        source_content = FileContentIdentity(
-                            file_object_id=stable_uuid(
-                                "file-identity",
-                                f"{system.hostname}:{source_path.casefold()}",
-                            ),
-                            version=1,
-                            size_bytes=content_rng.randint(600_000, 1_800_000),
-                            mime_type=(
-                                "application/gzip"
-                                if source_path.casefold().endswith((".gz", ".tgz"))
-                                else "application/octet-stream"
-                            ),
-                            seed_ref=stable_uuid(
-                                "storyline-scp-content",
-                                system.hostname,
-                                source_path,
-                            ),
-                        )
-                    orig_bytes = (
-                        source_content.size_bytes + rng.randint(8_192, 32_768)
-                        if source_content is not None
-                        else rng.randint(20_000, 250_000)
-                    )
-                    resp_bytes = rng.randint(4_000, 40_000)
-                    if (
-                        modeled_linux_receiver
-                        and target_system is not None
-                        and scp_destination is not None
-                    ):
-                        target_user = self._resolve_scp_target_user(
-                            extracted_username=scp_destination[2],
-                            fallback_username=process_actor.username,
-                        )
-                        transport_uid = self.activity_generator.generate_ssh_session(
-                            user=self.activity_generator._user_model_for_username(target_user),
-                            target_system=target_system,
-                            time=transfer_time,
-                            source_ip=system.ip,
-                            source_system=system,
-                            source_port=source_port,
-                            source_pid=pid,
-                            source_process_image=process_name,
-                            duration=transfer_duration,
-                            orig_bytes=orig_bytes,
-                            resp_bytes=resp_bytes,
-                            auth_method="publickey",
-                            emit_session_close=True,
-                            defer_session_close=True,
-                            source="storyline_scp",
-                        )
-                        network_plan_getter = getattr(
-                            self.dispatcher,
-                            "network_plan_for",
-                            None,
-                        )
-                        transport_plan = (
-                            network_plan_getter(transport_uid)
-                            if callable(network_plan_getter)
-                            else None
-                        )
-                        transfer_completed_at = (
-                            transport_plan.closed_at
-                            if transport_plan is not None
-                            else transfer_time + timedelta(seconds=transfer_duration)
-                        )
-                        self._emit_scp_receiver_artifacts(
-                            source_system=system,
-                            target_system=target_system,
-                            actor=process_actor,
-                            source_pid=pid,
-                            source_process=process_name,
-                            source_command=command_line,
-                            source_path=source_path,
-                            target_user=target_user,
-                            target_path=scp_destination[1],
-                            transfer_time=transfer_time,
-                            source_port=source_port,
-                            transfer_completed_at=transfer_completed_at,
-                            source_content=source_content,
-                            rng=rng,
-                        )
-                    else:
-                        self.activity_generator.generate_connection(
-                            src_ip=system.ip,
-                            dst_ip=dst_ip,
-                            time=transfer_time,
-                            dst_port=22,
-                            proto="tcp",
-                            service="ssh",
-                            duration=transfer_duration,
-                            orig_bytes=orig_bytes,
-                            resp_bytes=resp_bytes,
-                            conn_state="SF",
-                            emit_dns=not _is_private_ip(dst_ip),
-                            source_system=system,
-                            pid=pid,
-                            process_image=process_name,
-                            src_port=source_port,
-                        )
-
-            _EXPLICIT_CRED_TOOLS = {"psexec", "wmic", "runas", "schtasks"}
-            proc_basename = (
-                process_name.rsplit("\\", 1)[-1].lower()
-                if "\\" in process_name
-                else process_name.lower()
-            )
-            command_lower = command_line.lower()
-            uses_explicit_creds = proc_basename in _EXPLICIT_CRED_TOOLS or (
-                proc_basename in {"net.exe", "net1.exe"}
-                and any(token in command_lower for token in ("/user:", " /u:", " /user "))
-            )
-            if uses_explicit_creds and os_category == "windows":
-                cred_time = time - timedelta(milliseconds=rng.randint(5, 50))
-                self.activity_generator.generate_explicit_credentials(
-                    user=process_actor,
-                    system=system,
-                    time=cred_time,
-                    target_username=process_actor.username,
-                    target_server="localhost",
-                    process_name=process_name,
-                    process_pid=pid,
-                )
-
-            if os_category == "windows" and getattr(spec, "supplementary", "auto") != "none":
-                self.activity_generator._expand_and_emit(
-                    "process_create",
-                    time,
-                    actor=process_actor,
-                    target_system=system,
-                    command_line=command_line,
-                    os_category=os_category,
-                    source_pid=pid,
-                    logon_id=process_logon_id,
-                    skip_types=explicit_types,
-                )
-
-            # Mark as story process and schedule termination
-            self.state_manager.mark_story_process(system.hostname, pid)
-            lifetime = _estimate_process_lifetime(process_name, process_command_line)
-            if lifetime is not None:
-                term_delay = rng.uniform(lifetime[0], lifetime[1])
-                term_time = time + timedelta(seconds=term_delay)
-                shell_release_time = term_time
-                terminate_immediately = False
-                if os_category == "linux":
-                    from evidenceforge.generation.activity.generator import (
-                        _linux_foreground_lifetime,
-                    )
-
-                    terminate_immediately = (
-                        _linux_foreground_lifetime(process_name, process_command_line) is not None
-                    )
-                    if process_ref is not None:
-                        terminate_immediately = False
-                    if terminate_immediately and self._process_has_following_same_host_connection(
-                        system,
-                        future_specs,
-                    ):
-                        terminate_immediately = False
-                if terminate_immediately:
-                    self.activity_generator.generate_process_termination(
-                        user=process_actor,
-                        system=system,
-                        time=term_time,
-                        pid=pid,
-                        process_name=process_name,
-                        logon_id=process_logon_id,
-                        from_storyline=True,
-                    )
-                else:
-                    release_storyline_index = (
-                        self._storyline_process_ref_release_index(
-                            actor=process_actor,
-                            system=system,
-                            process_ref=process_ref,
-                        )
-                        if process_ref is not None
-                        else None
-                    )
-                    self._queue_story_process_termination(
-                        actor=process_actor,
-                        system=system,
-                        time=term_time,
-                        pid=pid,
-                        process_name=process_name,
-                        logon_id=process_logon_id,
-                        release_storyline_index=release_storyline_index,
-                    )
-                if os_category == "linux":
-                    self.activity_generator.remember_linux_foreground_process_completion(
-                        system=system,
-                        username=process_actor.username,
-                        logon_id=process_logon_id,
-                        parent_pid=parent_pid,
-                        termination_time=shell_release_time,
-                        process_name=process_name,
-                        command_line=process_command_line,
-                    )
-                    self._storyline_shell_available_at[shell_key] = shell_release_time
-                    process_shell_key = (system.hostname, process_actor.username)
-                    self._storyline_shell_available_at[process_shell_key] = shell_release_time
-
-        elif spec.type == "smb_activity":
-            time = self._storyline_smb_file_ready_time(
-                system=system,
-                spec=spec,
-                requested_at=time,
+                authored_time_shift=authored_time_shift,
+                session_required_until=session_required_until,
                 rng=rng,
-            )
-            malicious_event["time"] = time
-            smb_actor, smb_spec = self._storyline_smb_actor_and_spec(
-                actor,
-                system,
-                time,
-                spec,
-            )
-            result = self.activity_generator.generate_smb_activity(
-                spec=smb_spec,
-                actor=smb_actor,
-                parent_system=system,
-                time=time,
-                client_source_override=self._storyline_smb_source_override(
-                    system=system,
-                    spec=smb_spec,
-                ),
-            )
-            malicious_event.update(
-                {
-                    "session_id": result.session_id,
-                    "tree_ids": list(result.tree_ids),
-                    "transport_uids": list(result.transport_uids),
-                    "operations": list(result.operations),
-                    "batch_summary": {
-                        "selected": len(result.operations),
-                        "outcomes": sorted(
-                            {operation["outcome"] for operation in result.operations}
-                        ),
-                    },
-                }
-            )
-
-        elif spec.type == "connection":
-            source_ip = spec.source_ip or system.ip
-            dst_ip = spec.dst_ip
-            effective_dst_ip = dst_ip
-            if spec.hostname and not effective_dst_ip:
-                resolved_dst_ip = self._resolve_scenario_network_host(
-                    spec.hostname,
-                    src_host=system.hostname,
-                )
-                if resolved_dst_ip:
-                    effective_dst_ip = resolved_dst_ip
-            if not effective_dst_ip:
-                effective_dst_ip = self.public_identity_registry.bind(
-                    "c2",
-                    f"storyline-connection:{system.hostname}:{time.isoformat()}:{spec.dst_port}",
-                    forward_hostname=spec.hostname,
-                ).ip
-            if (
-                not _is_private_ip(source_ip)
-                and hasattr(self, "dispatcher")
-                and self.dispatcher.visibility_engine
-            ):
-                effective_dst_ip = self.dispatcher.visibility_engine._real_ip_to_vip.get(
-                    dst_ip, dst_ip
-                )
-            dst_port = spec.dst_port
-            service = spec.service or (
-                "ssl" if dst_port == 443 else "http" if dst_port == 80 else "ssl"
-            )
-            s_ob, s_rb = _size_storyline_connection(spec, rng)
-            # Build HttpContext if HTTP fields are provided
-            http_ctx = None
-            authored_request_body_len = getattr(spec, "request_body_len", None)
-            authored_request_multipart = getattr(spec, "request_multipart", None)
-            authored_response_multipart = getattr(spec, "response_multipart", None)
-            if (
-                spec.method
-                or spec.uri
-                or authored_request_body_len is not None
-                or authored_request_multipart is not None
-                or authored_response_multipart is not None
-            ):
-                from evidenceforge.events.contexts import HttpContext
-
-                # Context-aware response sizing (or author-specified override)
-                _method = spec.method or (
-                    "POST" if authored_request_body_len or authored_request_multipart else "GET"
-                )
-                _uri_raw = spec.uri or "/"
-                _mime_type = normalize_mime_type_for_path(_uri_raw, "text/html")
-                _is_c2_http = _is_c2_http_request(
-                    description=spec.description,
-                    technique=spec.technique,
-                    uri=_uri_raw,
-                    activity=activity,
-                )
-                if _is_c2_http and _mime_type == "text/html":
-                    _mime_type = rng.choices(
-                        ["application/json", "text/plain", "application/octet-stream"],
-                        weights=[55, 25, 20],
-                        k=1,
-                    )[0]
-                from evidenceforge.generation.activity.referrer import pick_referrer
-
-                _http_host = spec.hostname or effective_dst_ip
-                resp_bytes = _storyline_http_response_body_len(
-                    spec=spec,
-                    rng=rng,
-                    method=_method,
-                    uri=_uri_raw,
-                    host=_http_host,
-                    is_c2_http=_is_c2_http,
-                    use_connection_path_hints=True,
-                )
-                request_body_len = (
-                    authored_request_body_len
-                    if authored_request_body_len is not None
-                    else max(0, s_ob or 0)
-                    if _method not in {"GET", "HEAD", "CONNECT", "OPTIONS"}
-                    else 0
-                )
-                if request_body_len == 0 and _method == "POST":
-                    request_body_len = rng.randint(100, 10000)
-                http_ctx = HttpContext(
-                    method=_method,
-                    host=_http_host,
-                    uri=_uri_raw,
-                    version="1.1",
-                    user_agent=spec.user_agent or "Mozilla/5.0",
-                    request_body_len=request_body_len,
-                    response_body_len=resp_bytes,
-                    status_code=spec.status_code or 200,
-                    status_msg={
-                        200: "OK",
-                        301: "Moved Permanently",
-                        302: "Found",
-                        403: "Forbidden",
-                        404: "Not Found",
-                        500: "Internal Server Error",
-                    }.get(spec.status_code or 200, "OK"),
-                    referrer=spec.referrer
-                    if spec.referrer is not None
-                    else ""
-                    if _is_c2_http and rng.random() < 0.8
-                    else pick_referrer(rng, _http_host, context="general"),
-                    resp_mime_types=[_mime_type] if (spec.status_code or 200) == 200 else [],
-                    tags=[],
-                )
-                if (
-                    authored_request_multipart is not None
-                    or authored_response_multipart is not None
-                ):
-                    from evidenceforge.generation.activity.http_multipart import (
-                        apply_http_multipart_specs,
-                    )
-
-                    http_ctx = apply_http_multipart_specs(
-                        http_ctx,
-                        stable_key=(
-                            f"storyline:{system.hostname}:{source_ip}:{effective_dst_ip}:"
-                            f"{time.isoformat()}:{_uri_raw}"
-                        ),
-                        request_spec=authored_request_multipart,
-                        response_spec=authored_response_multipart,
-                        request_body_assertion=authored_request_body_len,
-                        response_body_assertion=spec.response_body_len,
-                    )
-
-            # Resolve source system from source_ip (not storyline system, which may be the target)
-            src_sys = None
-            ip_map = getattr(self.activity_generator, "_ip_to_system", {})
-            if source_ip in ip_map:
-                src_sys = ip_map[source_ip]
-            elif source_ip == system.ip:
-                src_sys = system
-            story_pid, story_image = self._last_storyline_process_for_system(src_sys)
-            story_proc = (
-                self.state_manager.get_process(src_sys.hostname, story_pid)
-                if story_pid > 0 and src_sys is not None
-                else None
-            )
-            story_command = story_proc.command_line if story_proc is not None else ""
-            if src_sys is not None and not story_command:
-                recorded_process = getattr(
-                    self, "_last_storyline_process_command_by_system", {}
-                ).get(src_sys.hostname)
-                if recorded_process is not None:
-                    recorded_pid, recorded_image, recorded_command = recorded_process
-                    target_url = f"{spec.hostname or effective_dst_ip}{spec.uri or '/'}"
-                    if target_url in recorded_command:
-                        story_pid = recorded_pid
-                        story_image = recorded_image
-                        story_command = recorded_command
-            upload_read_emitted = False
-            if http_ctx is not None and http_ctx.request_body_len > 0:
-                if http_ctx.request_multipart is not None:
-                    command_multipart = self._http_request_multipart_from_command(
-                        story_command,
-                        http_ctx.request_body_len,
-                        stable_key=(
-                            f"storyline-curl-validation:{source_ip}:{effective_dst_ip}:"
-                            f"{time.isoformat()}:{http_ctx.uri}"
-                        ),
-                    )
-                    if command_multipart is not None:
-                        self._validate_multipart_command_agreement(
-                            http_ctx.request_multipart,
-                            command_multipart,
-                        )
-                request_multipart = (
-                    None
-                    if http_ctx.request_multipart is not None
-                    else self._http_request_multipart_from_command(
-                        story_command,
-                        http_ctx.request_body_len,
-                        stable_key=(
-                            f"storyline-curl:{source_ip}:{effective_dst_ip}:"
-                            f"{time.isoformat()}:{http_ctx.uri}"
-                        ),
-                    )
-                )
-                request_entity = (
-                    None
-                    if request_multipart is not None or http_ctx.request_multipart is not None
-                    else self._http_request_entity_from_command(
-                        story_command, http_ctx.request_body_len
-                    )
-                )
-                if request_multipart is not None:
-                    http_ctx = replace(
-                        http_ctx,
-                        request_content_type=(
-                            f"{request_multipart.media_type}; boundary={request_multipart.boundary}"
-                        ),
-                        request_multipart=request_multipart,
-                    )
-                elif request_entity is not None:
-                    http_ctx = replace(
-                        http_ctx,
-                        request_content_type=request_entity.mime_type,
-                        request_entity=request_entity,
-                    )
-                    if not _is_exfil_connection_spec(spec):
-                        self._emit_http_upload_file_read(
-                            actor=actor,
-                            system=src_sys,
-                            pid=story_pid,
-                            process_image=story_image,
-                            command_line=story_command,
-                            entity=request_entity,
-                            connection_time=time,
-                        )
-                        upload_read_emitted = True
-            if story_pid > 0 and src_sys is not None and service in {"ssl", "https"}:
-                if self._command_contains_raw_tcp_endpoint(
-                    story_command,
-                    effective_dst_ip,
-                    dst_port,
-                ):
-                    service = ""
-            # Only use explicit hostname from scenario.  Do NOT fall back to
-            # Hostname resolution for storyline connections:
-            # - Explicit hostname → use it, emit DNS
-            # - No hostname but IP in REVERSE_DNS → use known hostname, emit DNS
-            # - No hostname, unknown IP → suppress (raw-IP C2/exfil), no DNS
-            from evidenceforge.generation.activity.network import REVERSE_DNS
-
-            if spec.hostname:
-                conn_hostname = spec.hostname
-                emit_dns = True
-            elif effective_dst_ip in REVERSE_DNS:
-                conn_hostname = None  # let generate_connection resolve via REVERSE_DNS
-                emit_dns = True
-            else:
-                conn_hostname = ""  # suppress — raw IP
-                emit_dns = False
-            s_conn_state = spec.conn_state or "SF"
-            story_command = story_command or story_image or ""
-            if _is_exfil_connection_spec(spec):
-                story_pid, story_image, story_command = (
-                    self._ensure_storyline_upload_process_for_exfil(
-                        actor=actor,
-                        system=src_sys,
-                        time=time,
-                        spec=spec,
-                        current_pid=story_pid,
-                        current_image=story_image,
-                        rng=rng,
-                    )
-                )
-                if http_ctx is not None:
-                    upload_user_agent = self._storyline_http_user_agent_for_process(
-                        system=src_sys,
-                        process_image=story_image,
-                        command_line=story_command,
-                        rng=rng,
-                    )
-                    if upload_user_agent and (
-                        not (spec.user_agent or "").strip()
-                        or (http_ctx.user_agent or "").strip().lower() == "mozilla/5.0"
-                    ):
-                        http_ctx = replace(http_ctx, user_agent=upload_user_agent)
-                    if (
-                        http_ctx.request_entity is None
-                        and http_ctx.request_multipart is None
-                        and http_ctx.request_body_len > 0
-                    ):
-                        request_multipart = self._http_request_multipart_from_command(
-                            story_command,
-                            http_ctx.request_body_len,
-                            stable_key=(
-                                f"storyline-exfil-curl:{source_ip}:{effective_dst_ip}:"
-                                f"{time.isoformat()}:{http_ctx.uri}"
-                            ),
-                        )
-                        request_entity = (
-                            None
-                            if request_multipart is not None
-                            else self._http_request_entity_from_command(
-                                story_command, http_ctx.request_body_len
-                            )
-                        )
-                        if request_multipart is not None:
-                            http_ctx = replace(
-                                http_ctx,
-                                request_content_type=(
-                                    f"{request_multipart.media_type}; "
-                                    f"boundary={request_multipart.boundary}"
-                                ),
-                                request_multipart=request_multipart,
-                            )
-                        elif request_entity is not None:
-                            http_ctx = replace(
-                                http_ctx,
-                                request_content_type=request_entity.mime_type,
-                                request_entity=request_entity,
-                            )
-                    staged_archive = self._matching_storyline_staged_archive_for_exfil(
-                        source_ip=source_ip,
-                        exfil_time=time,
-                    )
-                    if (
-                        http_ctx.request_entity is None
-                        and http_ctx.request_multipart is None
-                        and http_ctx.request_body_len > 0
-                        and staged_archive is not None
-                        and src_sys is not None
-                    ):
-                        from evidenceforge.events.contexts import HttpRequestEntityContext
-
-                        local_path = self._local_staging_path_for_archive(
-                            actor,
-                            src_sys,
-                            staged_archive.archive_path,
-                        )
-                        mime_type = infer_mime_type_from_path(
-                            local_path,
-                            "application/octet-stream",
-                        )
-                        request_entity = HttpRequestEntityContext(
-                            size=http_ctx.request_body_len,
-                            mime_type=mime_type,
-                            content_identity=(
-                                f"staged-upload:{src_sys.hostname}:{local_path}:"
-                                f"{http_ctx.request_body_len}"
-                            ),
-                            local_source_path=local_path,
-                            local_source_filename=(
-                                local_path.replace("\\", "/").rsplit("/", 1)[-1]
-                            ),
-                        )
-                        http_ctx = replace(
-                            http_ctx,
-                            request_content_type=mime_type,
-                            request_entity=request_entity,
-                        )
-                    if (
-                        not upload_read_emitted
-                        and http_ctx.request_entity is not None
-                        and staged_archive is None
-                    ):
-                        self._emit_http_upload_file_read(
-                            actor=actor,
-                            system=src_sys,
-                            pid=story_pid,
-                            process_image=story_image,
-                            command_line=story_command,
-                            entity=http_ctx.request_entity,
-                            connection_time=time,
-                        )
-                        upload_read_emitted = True
-                self._emit_storyline_archive_transfer_before_exfil(
-                    actor=actor,
-                    source_ip=source_ip,
-                    exfil_time=time,
-                    upload_bytes=s_ob,
-                    source_pid=story_pid,
-                    source_process=story_image or "",
-                    source_command=story_command,
-                    rng=rng,
-                )
-            connection_time = self._clamp_after_storyline_process_source_create(
-                system=src_sys,
-                pid=story_pid,
-                network_time=time,
-                rng=rng,
-            )
-            if http_ctx is not None and http_ctx.request_multipart is not None:
-                self._emit_http_multipart_file_reads(
-                    actor=actor,
-                    system=src_sys,
-                    pid=story_pid,
-                    process_image=story_image,
-                    command_line=story_command,
-                    multipart=http_ctx.request_multipart,
-                    connection_time=connection_time,
-                )
-                http_ctx = replace(
-                    http_ctx,
-                    request_multipart=replace(
-                        http_ctx.request_multipart,
-                        local_reads_emitted=True,
-                    ),
-                )
-            ids_alerts = _build_ids_alert_contexts(
-                getattr(spec, "ids_alerts", []),
-                time=connection_time,
-                src_ip=source_ip,
-                dst_ip=effective_dst_ip,
-                dst_port=dst_port,
-                proto="tcp",
-                rng=rng,
-                source="storyline_connection",
-            )
-            uid = self.activity_generator.generate_connection(
-                src_ip=source_ip,
-                dst_ip=effective_dst_ip,
-                time=connection_time,
-                dst_port=dst_port,
-                service=service,
-                duration=rng.uniform(1.0, 30.0),
-                orig_bytes=s_ob,
-                resp_bytes=s_rb,
-                conn_state=s_conn_state,
-                emit_dns=emit_dns,
-                source_system=src_sys,
-                http=http_ctx,
-                ids_alerts=ids_alerts,
-                pid=story_pid,
-                process_image=story_image,
-                hostname=conn_hostname,
-                preserve_dst_ip=bool(spec.hostname),
-                preserve_explicit_payload=(
-                    spec.orig_bytes is not None or spec.resp_bytes is not None
-                ),
-            )
-            logged_dst_ip = getattr(
-                self.activity_generator,
-                "_last_connection_effective_dst_ip",
-                effective_dst_ip,
-            )
-            malicious_event["dst_ip"] = logged_dst_ip
-            malicious_event["dst_port"] = dst_port
-            malicious_event["uid"] = _ground_truth_uid(uid, source_ip, logged_dst_ip)
-            if http_ctx is not None and http_ctx.request_entity is not None:
-                malicious_event["http_upload"] = {
-                    "request_body_len": http_ctx.request_body_len,
-                    "mime_type": http_ctx.request_entity.mime_type,
-                    "local_source_path": http_ctx.request_entity.local_source_path,
-                    "local_source_filename": http_ctx.request_entity.local_source_filename,
-                    "wire_filename": http_ctx.request_entity.wire_filename or None,
-                }
-            if http_ctx is not None and (
-                http_ctx.request_multipart is not None or http_ctx.response_multipart is not None
-            ):
-                last_transfers = getattr(
-                    self.activity_generator,
-                    "_last_connection_file_transfers",
-                    (),
-                )
-                multipart_fuids: dict[tuple[int, ...], list[str]] = {}
-                response_multipart_fuids: dict[tuple[int, ...], list[str]] = {}
-                for transfer in last_transfers:
-                    if not transfer.multipart_part_path:
-                        continue
-                    target = multipart_fuids if transfer.is_orig else response_multipart_fuids
-                    target.setdefault(transfer.multipart_part_path, []).append(transfer.fuid)
-
-                def _multipart_truth(entity: Any, *, response: bool) -> dict[str, Any]:
-                    fuid_map = response_multipart_fuids if response else multipart_fuids
-                    return {
-                        "body_len": entity.body_len,
-                        "media_type": entity.media_type,
-                        "boundary": entity.boundary,
-                        "parts": [
-                            {
-                                "path": list(part.path),
-                                "name": part.name,
-                                "decoded_size": part.decoded_size,
-                                "encoded_size": part.encoded_size,
-                                "local_source_path": part.local_source_path or None,
-                                "local_source_filename": part.local_source_filename or None,
-                                "wire_filename": part.wire_filename or None,
-                                "declared_mime_type": part.declared_content_type or None,
-                                "detected_mime_type": part.detected_mime_type or None,
-                                "transfer_encoding": part.transfer_encoding,
-                                "endpoint_read_owner_pid": (
-                                    story_pid
-                                    if not response and part.local_source_path and story_pid > 0
-                                    else None
-                                ),
-                                "fuids": fuid_map.get(part.path, []),
-                            }
-                            for part in entity.leaf_parts()
-                        ],
-                    }
-
-                malicious_event["http_multipart"] = {
-                    "request": (
-                        _multipart_truth(http_ctx.request_multipart, response=False)
-                        if http_ctx.request_multipart is not None
-                        else None
-                    ),
-                    "response": (
-                        _multipart_truth(http_ctx.response_multipart, response=True)
-                        if http_ctx.response_multipart is not None
-                        else None
-                    ),
-                }
-            if getattr(spec, "ids_alerts", []):
-                malicious_event["ids_alerts"] = _ids_attachment_ground_truth(
-                    getattr(spec, "ids_alerts", [])
-                )
-
-        elif spec.type == "ssh_session":
-            target = next(
-                (s for s in self.scenario.environment.systems if s.ip == system.ip), system
-            )
-            authored_ids_alerts = _build_ids_alert_contexts(
-                getattr(spec, "ids_alerts", []),
-                time=time,
-                src_ip=spec.source_ip or system.ip,
-                dst_ip=target.ip,
-                dst_port=22,
-                proto="tcp",
-                rng=rng,
-                source="storyline_ssh_session",
-            )
-            session_end_plan = self._session_end_plan_for_current_start()
-            if hasattr(self, "world_planner"):
-                source_system = (
-                    self.world_model.system_for_ip(spec.source_ip)
-                    if spec.source_ip and hasattr(self, "world_model")
-                    else None
-                )
-                result = self.world_planner.bootstrap_user_session(
-                    user=actor,
-                    target_system=target,
-                    time=time,
-                    rng=rng,
-                    session_kind="ssh",
-                    source_system=source_system,
-                    allow_existing=False,
-                    source_ip_override=spec.source_ip,
-                    storyline_protected=True,
-                    required_until=session_required_until,
-                    session_end_plan=session_end_plan,
-                    ids_alerts=authored_ids_alerts,
-                )
-            else:
-                source_ip = spec.source_ip or system.ip
-                uid = self.activity_generator.generate_ssh_session(
-                    user=actor,
-                    target_system=target,
-                    time=time,
-                    source_ip=source_ip,
-                    min_duration=(
-                        max(
-                            30.0,
-                            (session_required_until - time).total_seconds() + 30.0,
-                        )
-                        if session_required_until is not None
-                        else None
-                    ),
-                    emit_session_close=True,
-                    session_end_plan=session_end_plan,
-                    ids_alerts=authored_ids_alerts,
-                )
-                result = SimpleNamespace(network_uid=uid)
-            if getattr(result, "session", None) is not None:
-                self._record_storyline_logon(
-                    actor,
-                    target,
-                    result.session.logon_id,
-                    source_ip=result.session.source_ip,
-                )
-                self._record_storyline_session_ready(
-                    system=target,
-                    actor=actor,
-                    session=result.session,
-                    rng=rng,
-                )
-            malicious_event["dst_ip"] = system.ip
-            malicious_event["dst_port"] = 22
-            result_source_ip = (
-                result.session.source_ip
-                if getattr(result, "session", None) is not None
-                else spec.source_ip or system.ip
-            )
-            malicious_event["uid"] = _ground_truth_uid(
-                result.network_uid or "",
-                result_source_ip,
-                target.ip,
-            )
-            if getattr(spec, "ids_alerts", []):
-                malicious_event["ids_alerts"] = _ids_attachment_ground_truth(spec.ids_alerts)
-
-        elif spec.type == "rdp_session":
-            target = next(
-                (s for s in self.scenario.environment.systems if s.ip == system.ip), system
-            )
-            authored_ids_alerts = _build_ids_alert_contexts(
-                getattr(spec, "ids_alerts", []),
-                time=time,
-                src_ip=spec.source_ip or system.ip,
-                dst_ip=target.ip,
-                dst_port=3389,
-                proto="tcp",
-                rng=rng,
-                source="storyline_rdp_session",
-            )
-            if hasattr(self, "world_planner"):
-                source_system = (
-                    self.world_model.system_for_ip(spec.source_ip)
-                    if spec.source_ip and hasattr(self, "world_model")
-                    else None
-                )
-                result = self.world_planner.bootstrap_user_session(
-                    user=actor,
-                    target_system=target,
-                    time=time,
-                    rng=rng,
-                    session_kind="rdp",
-                    source_system=source_system,
-                    allow_existing=False,
-                    source_ip_override=spec.source_ip,
-                    storyline_protected=True,
-                    session_end_plan=self._authored_rdp_session_end_plan(),
-                    ids_alerts=authored_ids_alerts,
-                )
-            else:
-                source_ip = spec.source_ip or system.ip
-                uid = self.activity_generator.generate_rdp_session(
-                    user=actor,
-                    target_system=target,
-                    time=time,
-                    source_ip=source_ip,
-                    session_end_plan=self._authored_rdp_session_end_plan(),
-                    ids_alerts=authored_ids_alerts,
-                )
-                result = SimpleNamespace(network_uid=uid)
-            if getattr(result, "session", None) is not None:
-                malicious_event["actor"] = result.session.username
-                self._record_storyline_logon(
-                    actor,
-                    target,
-                    result.session.logon_id,
-                    source_ip=result.session.source_ip,
-                )
-                self._record_storyline_session_ready(
-                    system=target,
-                    actor=actor,
-                    session=result.session,
-                    rng=rng,
-                )
-            malicious_event["dst_ip"] = system.ip
-            malicious_event["dst_port"] = 3389
-            result_source_ip = (
-                result.session.source_ip
-                if getattr(result, "session", None) is not None
-                else spec.source_ip or system.ip
-            )
-            malicious_event["uid"] = _ground_truth_uid(
-                result.network_uid or "",
-                result_source_ip,
-                target.ip,
-            )
-            if getattr(spec, "ids_alerts", []):
-                malicious_event["ids_alerts"] = _ids_attachment_ground_truth(spec.ids_alerts)
-
-        elif spec.type == "account_created":
-            dc = next(
-                (s for s in self.scenario.environment.systems if s.type == "domain_controller"),
-                system,
-            )
-            target_sid = spec.target_sid or self._make_domain_sid()
-            effect_time = self._clamp_after_recent_storyline_process_source_create(
-                system=system,
-                event_time=time,
-                rng=rng,
-            )
-            self.activity_generator.generate_account_created(
-                actor=actor,
-                system=dc,
-                time=effect_time,
-                target_username=spec.target_username,
-                target_sid=target_sid,
-            )
-            # Store SID for later reuse by group_member_added, account_deleted,
-            # and any _get_sid() lookups (Windows event rendering).
-            self._created_account_sids[spec.target_username] = target_sid
-            self._created_account_effect_times[
-                self._account_create_lookup_key(dc, spec.target_username)
-            ] = effect_time
-            self._record_storyline_host_available_after(
-                system=dc,
-                actor=actor,
-                time=effect_time,
-                rng=rng,
-            )
-            if self._recent_storyline_account_create_command(dc, spec.target_username):
-                self._emit_storyline_account_password_followups(
-                    actor=actor,
-                    system=dc,
-                    time=effect_time,
-                    target_username=spec.target_username,
-                    target_sid=target_sid,
-                )
-            malicious_event["target_username"] = spec.target_username
-
-        elif spec.type == "account_deleted":
-            dc = next(
-                (s for s in self.scenario.environment.systems if s.type == "domain_controller"),
-                system,
-            )
-            target_sid = (
-                spec.target_sid
-                or self._created_account_sids.get(spec.target_username)
-                or self._make_domain_sid()
-            )
-            effect_time = self._clamp_after_recent_storyline_process_source_create(
-                system=system,
-                event_time=time,
-                rng=rng,
-            )
-            self.activity_generator.generate_account_deleted(
-                actor=actor,
-                system=dc,
-                time=effect_time,
-                target_username=spec.target_username,
-                target_sid=target_sid,
-                from_storyline=True,
-            )
-            malicious_event["target_username"] = spec.target_username
-
-        elif spec.type == "group_member_added":
-            dc = next(
-                (s for s in self.scenario.environment.systems if s.type == "domain_controller"),
-                system,
-            )
-            group_rid = 512 if "admin" in spec.group_name.lower() else rng.randint(1100, 9999)
-            group_sid = self._make_domain_sid(group_rid)
-            # Reuse SID from earlier account_created event, or generate new
-            member_sid = (
-                self._created_account_sids.get(spec.member_name)
-                or self.activity_generator.sid_registry.get(spec.member_name)
-                or self._make_domain_sid()
-            )
-            effect_time = self._clamp_after_recent_storyline_process_source_create(
-                system=dc,
-                event_time=time,
-                rng=rng,
-            )
-            account_created_at = self._created_account_effect_times.get(
-                self._account_create_lookup_key(dc, spec.member_name)
-            )
-            if account_created_at is not None and effect_time <= account_created_at:
-                effect_time = account_created_at + timedelta(milliseconds=rng.randint(180, 950))
-            self.activity_generator.generate_group_membership_change(
-                actor=actor,
-                system=dc,
-                time=effect_time,
-                action="add",
-                scope=spec.scope,
-                group_name=spec.group_name,
-                group_sid=group_sid,
-                member_username=spec.member_name,
-                member_sid=member_sid,
-            )
-            self._record_storyline_host_available_after(
-                system=dc,
-                actor=actor,
-                time=effect_time,
-                rng=rng,
-            )
-            malicious_event["group_name"] = spec.group_name
-            malicious_event["member_name"] = spec.member_name
-
-        elif spec.type == "service_installed":
-            effect_time = self._clamp_after_recent_storyline_process_source_create(
-                system=system,
-                event_time=time,
-                rng=rng,
-            )
-            source_ip = getattr(spec, "source_ip", None)
-            remote_source_system = self.world_model.system_for_ip(source_ip) if source_ip else None
-            service_lifecycle_group_id = self._storyline_remote_service_lifecycle_id(
-                system,
-                spec.service_name,
-            )
-            self.activity_generator.generate_service_installed(
-                user=actor,
-                system=system,
-                time=effect_time,
-                service_name=spec.service_name,
-                service_file_name=spec.service_file_name,
-                service_start_type=self._recent_storyline_service_start_type(
-                    system,
-                    spec.service_name,
-                ),
-                service_account=spec.service_account,
-                lifecycle_group_id=service_lifecycle_group_id,
-                remote_source_system=remote_source_system,
-            )
-            self._record_storyline_service_install(
-                system=system,
-                service_name=spec.service_name,
-                service_file_name=spec.service_file_name,
-                service_account=spec.service_account,
-                time=effect_time,
-                lifecycle_group_id=service_lifecycle_group_id,
-            )
-            malicious_event["service_name"] = spec.service_name
-            if spec.service_file_name:
-                malicious_event["service_file_name"] = spec.service_file_name
-
-        elif spec.type == "scheduled_task_created":
-            task_content = spec.task_content
-            source_command_line = self._recent_storyline_scheduled_task_command(
-                system,
-                spec.task_name,
-            )
-            if not task_content:
-                task_content = (
-                    f'<?xml version="1.0" encoding="UTF-16"?>\n'
-                    f'<Task version="1.2" xmlns="http://schemas.microsoft.com/windows/2004/02/mit/task">\n'
-                    f'  <Actions Context="Author">\n'
-                    f"    <Exec>\n"
-                    f"      <Command>C:\\Windows\\System32\\cmd.exe</Command>\n"
-                    f'      <Arguments>/c "{spec.task_name}"</Arguments>\n'
-                    f"    </Exec>\n"
-                    f"  </Actions>\n"
-                    f"</Task>"
-                )
-            elif not task_content.lstrip().startswith(("<?xml", "<Task")):
-                task_content = (
-                    f'<?xml version="1.0" encoding="UTF-16"?>\n'
-                    f'<Task version="1.2" xmlns="http://schemas.microsoft.com/windows/2004/02/mit/task">\n'
-                    f'  <Actions Context="Author">\n'
-                    f"    <Exec>\n"
-                    f"      <Command>{task_content}</Command>\n"
-                    f"    </Exec>\n"
-                    f"  </Actions>\n"
-                    f"</Task>"
-                )
-            effect_time = self._clamp_after_recent_storyline_process_source_create(
-                system=system,
-                event_time=time,
-                rng=rng,
-            )
-            self.activity_generator.generate_scheduled_task(
-                user=actor,
-                system=system,
-                time=effect_time,
-                task_name=spec.task_name,
-                action="created",
-                task_content=task_content,
-                source_command_line=source_command_line,
-            )
-            malicious_event["task_name"] = spec.task_name
-            malicious_event["task_content"] = task_content
-
-        elif spec.type == "log_cleared":
-            effect_time = self._clamp_after_recent_storyline_process_source_create(
-                system=system,
-                event_time=time,
-                rng=rng,
-            )
-            subject_logon_id = self._recent_storyline_process_logon_id(
-                actor,
-                system,
-                effect_time,
-                executable="wevtutil.exe",
-            )
-            self.activity_generator.generate_log_cleared(
-                user=actor,
-                system=system,
-                time=effect_time,
-                from_storyline=True,
-                subject_logon_id=subject_logon_id,
-            )
-
-        elif spec.type == "create_remote_thread":
-            source_pid, source_image = self._last_storyline_process_for_system(
-                system,
-                actor=actor,
-            )
-            # Use a realistic target PID — look up the process name from
-            # system PIDs or use a plausible default (not 4 = System kernel)
-            target_image = _normalize_storyline_process_image(
-                spec.target_process,
-                _get_os_category(system.os),
-                username=actor.username,
-            )
-            target_name = target_image.rsplit("\\", 1)[-1].rsplit("/", 1)[-1].lower()
-            if source_pid <= 0:
-                # Without a live source process, there is no realistic Sysmon
-                # Event 8 relationship to render. Keep the storyline record,
-                # but mark it skipped instead of claiming generated evidence.
-                malicious_event["target_process"] = target_image
-                malicious_event["skipped_reason"] = "no_live_source_process"
-            else:
-                effect_time = self._clamp_after_storyline_process_source_create(
-                    system=system,
-                    pid=source_pid,
-                    network_time=time,
-                    rng=rng,
-                )
-                target_pid = self.activity_generator._get_system_pid(
-                    system.hostname,
-                    target_name.replace(".exe", ""),
-                    0x27C,  # 636 default
-                )
-                evidence_emitted = self.activity_generator.generate_create_remote_thread(
-                    user=actor,
-                    system=system,
-                    time=effect_time,
-                    source_pid=source_pid,
-                    source_image=source_image,
-                    target_pid=target_pid,
-                    target_image=target_image,
-                )
-                malicious_event["target_process"] = target_image
-                if not evidence_emitted:
-                    malicious_event["skipped_reason"] = "no_live_target_process"
-
-        elif spec.type == "process_access":
-            source_pid, source_image = self._last_storyline_process_for_system(
-                system,
-                actor=actor,
-            )
-            os_category = _get_os_category(system.os)
-            target_image = _normalize_storyline_process_image(
-                spec.target_process,
-                os_category,
-                username=actor.username,
-            )
-            if source_pid <= 0:
-                # Without a live source process, there is no realistic Sysmon
-                # Event 10 relationship to render. Keep the storyline record,
-                # but mark it skipped instead of claiming generated evidence.
-                malicious_event["target_process"] = target_image
-                malicious_event["skipped_reason"] = "no_live_source_process"
-            else:
-                target_name = target_image.rsplit("\\", 1)[-1].rsplit("/", 1)[-1].lower()
-                target_pid = self.activity_generator._get_system_pid(
-                    system.hostname,
-                    target_name.replace(".exe", ""),
-                    0x27C,
-                )
-                evidence_emitted = self.activity_generator.generate_process_access(
-                    user=actor,
-                    system=system,
-                    time=self._clamp_after_storyline_process_source_create(
-                        system=system,
-                        pid=source_pid,
-                        network_time=time,
-                        rng=rng,
-                    ),
-                    source_pid=source_pid,
-                    source_image=source_image,
-                    target_pid=target_pid,
-                    target_image=target_image,
-                    granted_access=spec.access_mask,
-                )
-                malicious_event["target_process"] = target_image
-                if not evidence_emitted:
-                    malicious_event["skipped_reason"] = "no_live_target_process"
-
-        elif spec.type == "dhcp_lease":
-            existing_lease = getattr(self, "_dhcp_lease_state", {}).get(system.hostname)
-            if spec.mac_address:
-                mac = spec.mac_address
-            elif existing_lease:
-                mac = existing_lease["mac"]
-            else:
-                ip_hash = _stable_seed(f"mac_{spec.requested_ip or system.ip}")
-                mac = (
-                    f"00:50:56:{(ip_hash >> 16) & 0xFF:02x}"
-                    f":{(ip_hash >> 8) & 0xFF:02x}:{ip_hash & 0xFF:02x}"
-                )
-            from evidenceforge.generation.world_model import HostCapability
-            from evidenceforge.utils.ids import generate_zeek_uid
-
-            dhcp_servers = self.world_model.systems_with_capability(
-                HostCapability.DHCP_SERVER,
-                distinct_from=system,
-            )
-            if not dhcp_servers:
-                raise StateError(
-                    f"DHCP lease for {system.hostname} requires a distinct modeled DHCP server"
-                )
-            dhcp_server = dhcp_servers[
-                _stable_seed(
-                    "storyline_dhcp_server:"
-                    f"{getattr(self, '_current_storyline_spec_id', '')}:{system.hostname}"
-                )
-                % len(dhcp_servers)
-            ].ip
-            lease_time = (
-                float(existing_lease["lease_time"])
-                if existing_lease
-                else float(rng.choice([3600, 7200, 14400, 86400]))
-            )
-            if existing_lease:
-                legacy_timer_state = "renewal_rng" not in existing_lease
-                renewal_sequence = int(existing_lease.get("renewal_sequence", 0))
-                renewal_rng = existing_lease.get("renewal_rng")
-                if renewal_rng is None:
-                    renewal_rng = random.Random(
-                        _stable_seed(f"dhcp_renewal_timer:{system.hostname}:{mac}")
-                    )
-                if "timer_granularity" in existing_lease:
-                    timer_granularity = float(existing_lease["timer_granularity"])
-                else:
-                    timer_granularity = renewal_rng.choice([0.25, 1.0, 1.0, 5.0])
-            else:
-                legacy_timer_state = False
-                renewal_sequence = 0
-                renewal_rng = random.Random(
-                    _stable_seed(f"dhcp_renewal_timer:{system.hostname}:{mac}")
-                )
-                timer_granularity = renewal_rng.choice([0.25, 1.0, 1.0, 5.0])
-            if legacy_timer_state:
-                renewal_interval = float(existing_lease["renewal_interval"])
-            else:
-                renewal_interval = dhcp_renewal_interval_seconds(
-                    lease_time,
-                    timing_runtime=self.timing_runtime,
-                    stable_id=f"{system.hostname}|{mac}",
-                    host=system.hostname,
-                    renewal_sequence=renewal_sequence,
-                    timer_granularity=timer_granularity,
-                )
-                renewal_sequence += 1
-            msg_types = ["REQUEST", "ACK"] if existing_lease else None
-            authored_ids_alerts = _build_ids_alert_contexts(
-                getattr(spec, "ids_alerts", []),
-                time=time,
-                src_ip=system.ip,
-                dst_ip=dhcp_server,
-                dst_port=67,
-                proto="udp",
-                rng=rng,
-                source="storyline_dhcp_lease",
-            )
-            self.activity_generator.generate_dhcp_lease(
-                system=system,
-                time=time,
-                mac=mac,
-                server_addr=dhcp_server,
-                lease_time=lease_time,
-                uid=generate_zeek_uid("C"),
-                msg_types=msg_types,
-                renewal_interval=renewal_interval,
-                ids_alerts=authored_ids_alerts,
-            )
-            if hasattr(self, "_dhcp_lease_state"):
-                self._dhcp_lease_state[system.hostname] = {
-                    "mac": mac,
-                    "lease_time": lease_time,
-                    "last_renewal": time.timestamp(),
-                    "next_renewal": time.timestamp() + renewal_interval,
-                    "renewal_interval": renewal_interval,
-                    "renewal_rng": renewal_rng,
-                    "renewal_sequence": renewal_sequence,
-                    "timer_granularity": timer_granularity,
-                    "server_addr": dhcp_server,
-                    "system": system,
-                }
-            malicious_event["mac_address"] = mac
-            if getattr(spec, "ids_alerts", []):
-                malicious_event["ids_alerts"] = _ids_attachment_ground_truth(spec.ids_alerts)
-
-        elif spec.type == "port_scan":
-            malicious_event = PortScanActionBundle(
-                executor=self,
-                request=PortScanRequest(
-                    spec=spec,
-                    actor=actor,
-                    system=system,
-                    time=time,
-                    rng=rng,
-                    malicious_event=malicious_event,
-                ),
-            ).execute()
-
-        elif spec.type == "beacon":
-            # Resolve timing parameters
-            start = self._parse_storyline_time(spec.start_time) if spec.start_time else time
-            interval_sec = parse_duration(spec.interval).total_seconds()
-            duration_sec = None
-            count = spec.count
-            if spec.duration is not None:
-                duration_sec = parse_duration(spec.duration).total_seconds()
-            elif spec.end_time is not None:
-                end_dt = self._parse_storyline_time(spec.end_time)
-                duration_sec = (end_dt - start).total_seconds()
-
-            beacon_profile: dict[str, Any] | None = None
-            profile_http_sequence: list[dict[str, Any]] = []
-            profile_user_agents: list[str] = []
-            profile_dns_resolution = None
-            if spec.profile:
-                from evidenceforge.config.beacon_profiles import get_profile
-
-                beacon_profile = get_profile(spec.profile)
-                if beacon_profile is None:
-                    raise ValueError(f"Unknown beacon profile: {spec.profile}")
-                raw_profile_sequence = beacon_profile.get("http_sequence", [])
-                if isinstance(raw_profile_sequence, list):
-                    profile_http_sequence = [
-                        entry for entry in raw_profile_sequence if isinstance(entry, dict)
-                    ]
-                raw_profile_user_agents = beacon_profile.get("user_agents", [])
-                if isinstance(raw_profile_user_agents, list):
-                    profile_user_agents = [str(value) for value in raw_profile_user_agents if value]
-                profile_dns_resolution = beacon_profile.get("dns_resolution")
-            explicit_http_sequence: list[BeaconHttpSequenceEntry] = list(spec.http_sequence)
-
-            beacon_src_ip = spec.source_ip or system.ip
-            beacon_dst_ip = spec.dst_ip
-            if spec.hostname and not beacon_dst_ip:
-                resolved_beacon_ip = self._resolve_scenario_network_host(
-                    spec.hostname,
-                    src_host=system.hostname,
-                )
-                if resolved_beacon_ip:
-                    beacon_dst_ip = resolved_beacon_ip
-
-            # Deny mode: firewall context
-            fw_ctx = None
-            deny_conn_state = None
-            if spec.action == "deny":
-                from evidenceforge.events.contexts import FirewallContext
-
-                deny_conn_state = self._get_firewall_deny_conn_state()
-                src_iface = self._resolve_firewall_interface(beacon_src_ip)
-                dst_iface = self._resolve_firewall_interface(beacon_dst_ip)
-                fw_ctx = FirewallContext(
-                    action="deny",
-                    msg_id=106023,
-                    connection_id=0,
-                    src_interface=src_iface,
-                    dst_interface=dst_iface,
-                    access_group=f"{src_iface}_access_in",
-                )
-
-            # Allow mode: resolve service, http context, hostname, byte sizing
-            service = spec.service
-            http_ctx = None
-            http_is_c2 = False
-            http_method = ""
-            http_uri = ""
-            conn_hostname = None
-            emit_dns = False
-            s_ob, s_rb = _size_storyline_connection(spec, rng)
-            s_conn_state = spec.conn_state or "SF"
-
-            if spec.action == "allow":
-                service = service or (
-                    "ssl" if spec.dst_port == 443 else "http" if spec.dst_port == 80 else "ssl"
-                )
-                # Build HttpContext if HTTP/proxy-visible request metadata is provided.
-                # HTTPS CONNECT beacons still need this for proxy User-Agent fidelity
-                # even though no origin-side Zeek http.log is emitted for TLS.
-                first_sequence_entry: BeaconHttpSequenceEntry | dict[str, Any] | None = None
-                if explicit_http_sequence:
-                    first_sequence_entry = explicit_http_sequence[0]
-                elif profile_http_sequence:
-                    first_sequence_entry = profile_http_sequence[0]
-                profile_user_agent = profile_user_agents[0] if profile_user_agents else None
-
-                if (
-                    spec.method
-                    or spec.uri
-                    or spec.user_agent
-                    or getattr(spec, "request_body_len", None) is not None
-                    or spec.request_multipart is not None
-                    or spec.response_multipart is not None
-                    or profile_user_agent is not None
-                    or first_sequence_entry is not None
-                ):
-                    from evidenceforge.events.contexts import HttpContext
-
-                    _method = _entry_value(first_sequence_entry, "method") or spec.method or "GET"
-                    _uri_template = _entry_value(first_sequence_entry, "uri") or spec.uri or "/"
-                    _uri_raw = _render_beacon_template(
-                        str(_uri_template),
-                        spec=spec,
-                        system=system,
-                        tick_index=0,
-                    )
-                    _user_agent = (
-                        _entry_value(first_sequence_entry, "user_agent")
-                        or spec.user_agent
-                        or profile_user_agent
-                        or "Mozilla/5.0"
-                    )
-                    http_method = _method
-                    http_uri = _uri_raw
-                    _mime_type = normalize_mime_type_for_path(_uri_raw, "text/html")
-                    _is_c2_http = _is_c2_http_request(
-                        description=spec.description,
-                        technique=spec.technique,
-                        uri=_uri_raw,
-                        activity=activity,
-                    )
-                    http_is_c2 = _is_c2_http
-                    if _is_c2_http and _mime_type == "text/html":
-                        _mime_type = rng.choices(
-                            ["application/json", "text/plain", "application/octet-stream"],
-                            weights=[65, 25, 10],
-                            k=1,
-                        )[0]
-                    from evidenceforge.generation.activity.referrer import pick_referrer
-
-                    _http_host2 = spec.hostname or spec.dst_ip
-                    response_override = _range_or_value(
-                        _entry_value(first_sequence_entry, "response_body_len"),
-                        rng,
-                    )
-                    resp_bytes = (
-                        response_override
-                        if response_override is not None
-                        else _storyline_http_response_body_len(
-                            spec=spec,
-                            rng=rng,
-                            method=_method,
-                            uri=_uri_raw,
-                            host=_http_host2,
-                            is_c2_http=_is_c2_http,
-                            use_connection_path_hints=False,
-                        )
-                    )
-                    request_override = _range_or_value(
-                        _entry_value(first_sequence_entry, "request_body_len"), rng
-                    )
-                    request_body_len = (
-                        request_override
-                        if request_override is not None
-                        else getattr(spec, "request_body_len", None)
-                        if getattr(spec, "request_body_len", None) is not None
-                        else max(0, s_ob or 0)
-                        if _method not in {"GET", "HEAD", "CONNECT", "OPTIONS"}
-                        else 0
-                    )
-                    if request_body_len == 0 and _method == "POST":
-                        request_body_len = rng.randint(100, 10000)
-                    _status_code = (
-                        _entry_value(first_sequence_entry, "status_code") or spec.status_code or 200
-                    )
-                    http_ctx = HttpContext(
-                        method=_method,
-                        host=_http_host2,
-                        uri=_uri_raw,
-                        version="1.1",
-                        user_agent=str(_user_agent),
-                        request_body_len=request_body_len,
-                        response_body_len=resp_bytes,
-                        status_code=_status_code,
-                        status_msg={
-                            200: "OK",
-                            301: "Moved Permanently",
-                            302: "Found",
-                            403: "Forbidden",
-                            404: "Not Found",
-                            500: "Internal Server Error",
-                        }.get(_status_code, "OK"),
-                        referrer=spec.referrer
-                        if spec.referrer is not None
-                        else ""
-                        if _is_c2_http and rng.random() < 0.75
-                        else pick_referrer(rng, _http_host2, context="general"),
-                        resp_mime_types=[_mime_type] if _status_code == 200 else [],
-                        tags=[],
-                    )
-                    first_request_multipart = (
-                        _entry_value(first_sequence_entry, "request_multipart")
-                        or spec.request_multipart
-                    )
-                    first_response_multipart = (
-                        _entry_value(first_sequence_entry, "response_multipart")
-                        or spec.response_multipart
-                    )
-                    if first_request_multipart is not None or first_response_multipart is not None:
-                        from evidenceforge.generation.activity.http_multipart import (
-                            apply_http_multipart_specs,
-                        )
-
-                        http_ctx = apply_http_multipart_specs(
-                            http_ctx,
-                            stable_key=(
-                                f"beacon:{system.hostname}:{beacon_src_ip}:{beacon_dst_ip}:"
-                                f"{start.isoformat()}:0:{_uri_raw}"
-                            ),
-                            request_spec=first_request_multipart,
-                            response_spec=first_response_multipart,
-                            request_body_assertion=request_override,
-                            response_body_assertion=response_override,
-                        )
-
-                # Hostname / DNS resolution (same logic as connection handler)
-                from evidenceforge.generation.activity.network import REVERSE_DNS
-
-                if spec.hostname:
-                    conn_hostname = spec.hostname
-                    emit_dns = True
-                elif spec.dst_ip in REVERSE_DNS:
-                    conn_hostname = None
-                    emit_dns = True
-                else:
-                    conn_hostname = ""
-                    emit_dns = False
-
-            # Resolve source system
-            src_sys = None
-            ip_map = getattr(self.activity_generator, "_ip_to_system", {})
-            if beacon_src_ip in ip_map:
-                src_sys = ip_map[beacon_src_ip]
-            elif beacon_src_ip == system.ip:
-                src_sys = system
-            story_pid, story_image = self._last_storyline_process_for_system(src_sys)
-
-            attempt_count = 0
-            for tick_time in _iter_periodic_ticks(
-                start,
-                interval_sec,
-                duration_sec,
-                count,
-                spec.jitter,
-                rng,
-                exclusive_end_time=getattr(self, "end_time", None),
-            ):
-                self.state_manager.set_current_time(tick_time)
-                tick_sequence_entry: BeaconHttpSequenceEntry | dict[str, Any] | None = None
-                if explicit_http_sequence:
-                    tick_sequence_entry = explicit_http_sequence[
-                        attempt_count % len(explicit_http_sequence)
-                    ]
-                elif profile_http_sequence:
-                    tick_sequence_entry = _weighted_profile_entry(
-                        profile_http_sequence,
-                        tick_index=attempt_count,
-                        spec=spec,
-                        system=system,
-                    )
-                tick_method = _entry_value(tick_sequence_entry, "method") or spec.method or "GET"
-                tick_uri_template = _entry_value(tick_sequence_entry, "uri") or spec.uri or "/"
-                tick_uri = _render_beacon_template(
-                    str(tick_uri_template),
-                    spec=spec,
-                    system=system,
-                    tick_index=attempt_count,
-                )
-                tick_user_agent = (
-                    _entry_value(tick_sequence_entry, "user_agent")
-                    or spec.user_agent
-                    or (
-                        profile_user_agents[
-                            _stable_seed(
-                                f"beacon-ua:{system.hostname}:{spec.profile}:{attempt_count}"
-                            )
-                            % len(profile_user_agents)
-                        ]
-                        if profile_user_agents
-                        else None
-                    )
-                    or "Mozilla/5.0"
-                )
-                tick_status_code = (
-                    _entry_value(tick_sequence_entry, "status_code") or spec.status_code or 200
-                )
-                tick_response_override = _range_or_value(
-                    _entry_value(tick_sequence_entry, "response_body_len"),
-                    rng,
-                )
-                tick_request_override = _range_or_value(
-                    _entry_value(tick_sequence_entry, "request_body_len"), rng
-                )
-                tick_orig_bytes = (
-                    _range_or_value(_entry_value(tick_sequence_entry, "orig_bytes"), rng)
-                    if tick_sequence_entry is not None
-                    else None
-                )
-                if tick_orig_bytes is None:
-                    tick_orig_bytes = s_ob
-                tick_resp_bytes = (
-                    _range_or_value(_entry_value(tick_sequence_entry, "resp_bytes"), rng)
-                    if tick_sequence_entry is not None
-                    else None
-                )
-                if tick_resp_bytes is None:
-                    tick_resp_bytes = s_rb
-                tick_emit_dns = emit_dns and (
-                    spec.dns_resolution == "each_tick"
-                    or (profile_dns_resolution == "each_tick" and spec.dns_resolution == "cached")
-                    or attempt_count == 0
-                )
-                tick_http_ctx = http_ctx
-                if tick_http_ctx is not None and (
-                    tick_sequence_entry is not None
-                    or spec.request_multipart is not None
-                    or spec.response_multipart is not None
-                ):
-                    tick_mime_type = normalize_mime_type_for_path(tick_uri, "text/html")
-                    tick_request_body_len = (
-                        tick_request_override
-                        if tick_request_override is not None
-                        else getattr(spec, "request_body_len", None)
-                        if getattr(spec, "request_body_len", None) is not None
-                        else max(0, tick_orig_bytes)
-                        if str(tick_method).upper() not in {"GET", "HEAD", "CONNECT", "OPTIONS"}
-                        else 0
-                    )
-                    tick_response_body_len = (
-                        tick_response_override
-                        if tick_response_override is not None
-                        else tick_http_ctx.response_body_len
-                    )
-                    tick_http_ctx = replace(
-                        tick_http_ctx,
-                        method=str(tick_method),
-                        uri=tick_uri,
-                        user_agent=str(tick_user_agent),
-                        status_code=int(tick_status_code),
-                        status_msg={
-                            200: "OK",
-                            301: "Moved Permanently",
-                            302: "Found",
-                            403: "Forbidden",
-                            404: "Not Found",
-                            500: "Internal Server Error",
-                        }.get(int(tick_status_code), "OK"),
-                        response_body_len=tick_response_body_len,
-                        response_multipart=None,
-                        request_body_len=tick_request_body_len,
-                        request_entity=None,
-                        request_multipart=None,
-                        referrer=_entry_value(tick_sequence_entry, "referrer")
-                        if tick_sequence_entry is not None
-                        and _entry_value(tick_sequence_entry, "referrer") is not None
-                        else tick_http_ctx.referrer,
-                        tags=list(tick_http_ctx.tags),
-                        resp_fuids=list(tick_http_ctx.resp_fuids),
-                        resp_mime_types=[tick_mime_type] if int(tick_status_code) == 200 else [],
-                    )
-                    tick_request_multipart = (
-                        _entry_value(tick_sequence_entry, "request_multipart")
-                        or spec.request_multipart
-                    )
-                    tick_response_multipart = (
-                        _entry_value(tick_sequence_entry, "response_multipart")
-                        or spec.response_multipart
-                    )
-                    if tick_request_multipart is not None or tick_response_multipart is not None:
-                        from evidenceforge.generation.activity.http_multipart import (
-                            apply_http_multipart_specs,
-                        )
-
-                        tick_http_ctx = apply_http_multipart_specs(
-                            tick_http_ctx,
-                            stable_key=(
-                                f"beacon:{system.hostname}:{beacon_src_ip}:{beacon_dst_ip}:"
-                                f"{tick_time.isoformat()}:{attempt_count}:{tick_uri}"
-                            ),
-                            request_spec=tick_request_multipart,
-                            response_spec=tick_response_multipart,
-                            request_body_assertion=tick_request_override,
-                            response_body_assertion=tick_response_override,
-                        )
-                    if tick_response_override is not None:
-                        tick_resp_bytes = max(
-                            tick_resp_bytes,
-                            tick_response_override + rng.randint(300, 5000),
-                        )
-                if (
-                    http_ctx is not None
-                    and http_is_c2
-                    and spec.response_body_len is None
-                    and tick_response_override is None
-                    and (tick_http_ctx or http_ctx).response_multipart is None
-                ):
-                    tick_http_body_len = _c2_http_response_size(
-                        rng,
-                        method=str(tick_method or http_method or http_ctx.method),
-                        uri=tick_uri or http_uri or http_ctx.uri,
-                    )
-                    tick_http_ctx = replace(
-                        tick_http_ctx or http_ctx,
-                        response_body_len=tick_http_body_len,
-                        tags=list((tick_http_ctx or http_ctx).tags),
-                        resp_fuids=list((tick_http_ctx or http_ctx).resp_fuids),
-                        resp_mime_types=list((tick_http_ctx or http_ctx).resp_mime_types),
-                    )
-                    tick_resp_bytes = max(
-                        tick_resp_bytes,
-                        tick_http_body_len + rng.randint(300, 5000),
-                    )
-                if story_pid <= 0:
-                    story_pid, story_image = self._ensure_storyline_service_process_for_beacon(
-                        actor,
-                        src_sys,
-                        tick_time,
-                    )
-                if tick_http_ctx is not None and tick_http_ctx.request_multipart is not None:
-                    running_beacon = (
-                        self.state_manager.get_process(src_sys.hostname, story_pid)
-                        if src_sys is not None and story_pid > 0
-                        else None
-                    )
-                    self._emit_http_multipart_file_reads(
-                        actor=actor,
-                        system=src_sys,
-                        pid=story_pid,
-                        process_image=story_image,
-                        command_line=(
-                            running_beacon.command_line if running_beacon is not None else ""
-                        ),
-                        multipart=tick_http_ctx.request_multipart,
-                        connection_time=tick_time,
-                    )
-                    tick_http_ctx = replace(
-                        tick_http_ctx,
-                        request_multipart=replace(
-                            tick_http_ctx.request_multipart,
-                            local_reads_emitted=True,
-                        ),
-                    )
-                tick_ids_alerts = _build_ids_alert_contexts(
-                    getattr(spec, "ids_alerts", []),
-                    time=tick_time,
-                    src_ip=beacon_src_ip,
-                    dst_ip=beacon_dst_ip,
-                    dst_port=spec.dst_port,
-                    proto=spec.protocol,
-                    rng=rng,
-                    source="storyline_beacon",
-                )
-                if spec.action == "deny":
-                    proxy_chain = getattr(self.activity_generator, "_proxy_routes", {}).get(
-                        beacon_src_ip
-                    )
-                    explicit_proxy = (
-                        getattr(self.activity_generator, "_proxy_mode", "transparent") == "explicit"
-                        and proxy_chain
-                        and spec.protocol == "tcp"
-                        and spec.dst_port in (80, 443)
-                    )
-                    if explicit_proxy:
-                        from evidenceforge.events.contexts import ProxyContext
-
-                        proxy_sys = proxy_chain[0]
-                        beacon_host = spec.hostname or spec.dst_ip
-                        proxy_method = "CONNECT" if spec.dst_port == 443 else str(tick_method)
-                        proxy_url = (
-                            f"{beacon_host}:443"
-                            if proxy_method == "CONNECT"
-                            else f"http://{beacon_host}{tick_uri}"
-                        )
-                        proxy_user_agent = str(tick_user_agent)
-                        proxy_source_system = getattr(
-                            self.activity_generator,
-                            "_ip_to_system",
-                            {},
-                        ).get(beacon_src_ip)
-                        proxy_ctx = ProxyContext(
-                            client_ip=beacon_src_ip,
-                            username=self.activity_generator._proxy_username_for_source(
-                                source_system=proxy_source_system,
-                                user_agent=proxy_user_agent,
-                                cache_result="DENIED",
-                                hostname=beacon_host,
-                                time=tick_time,
-                            ),
-                            method=proxy_method,
-                            url=proxy_url,
-                            host=beacon_host,
-                            status_code=403,
-                            sc_bytes=rng.randint(500, 2000),
-                            cs_bytes=rng.randint(180, 520),
-                            time_taken=rng.randint(20, 1500),
-                            user_agent=proxy_user_agent,
-                            content_type="text/html",
-                            cache_result="DENIED",
-                            referrer=spec.referrer or "",
-                            proxy_fqdn=self.activity_generator._proxy_fqdn(proxy_sys),
-                            proxy_action="deny",
-                        )
-                        self.activity_generator.generate_connection(
-                            src_ip=beacon_src_ip,
-                            dst_ip=beacon_dst_ip,
-                            time=tick_time,
-                            dst_port=spec.dst_port,
-                            proto=spec.protocol,
-                            service="ssl" if spec.dst_port == 443 else "http",
-                            duration=rng.uniform(0.05, 2.0),
-                            orig_bytes=tick_orig_bytes,
-                            resp_bytes=tick_resp_bytes,
-                            conn_state="SF",
-                            emit_dns=tick_emit_dns,
-                            source_system=src_sys,
-                            http=tick_http_ctx,
-                            ids_alerts=tick_ids_alerts,
-                            proxy=proxy_ctx,
-                            hostname=conn_hostname if conn_hostname is not None else spec.hostname,
-                            pid=story_pid,
-                            process_image=story_image,
-                            preserve_dst_ip=bool(spec.hostname),
-                            preserve_explicit_payload=(
-                                spec.orig_bytes is not None
-                                or spec.resp_bytes is not None
-                                or tick_sequence_entry is not None
-                            ),
-                        )
-                    else:
-                        self.activity_generator.generate_connection(
-                            src_ip=beacon_src_ip,
-                            dst_ip=beacon_dst_ip,
-                            time=tick_time,
-                            dst_port=spec.dst_port,
-                            proto=spec.protocol,
-                            conn_state=deny_conn_state,
-                            firewall=fw_ctx,
-                            ids_alerts=tick_ids_alerts,
-                            emit_dns=False,
-                        )
-                else:
-                    # Allow DNS only on the first tick; cache handles the rest
-                    self.activity_generator.generate_connection(
-                        src_ip=beacon_src_ip,
-                        dst_ip=beacon_dst_ip,
-                        time=tick_time,
-                        dst_port=spec.dst_port,
-                        proto=spec.protocol,
-                        service=service,
-                        duration=rng.uniform(0.5, 10.0),
-                        orig_bytes=tick_orig_bytes,
-                        resp_bytes=tick_resp_bytes,
-                        conn_state=s_conn_state,
-                        emit_dns=tick_emit_dns,
-                        source_system=src_sys,
-                        http=tick_http_ctx,
-                        ids_alerts=tick_ids_alerts,
-                        hostname=conn_hostname,
-                        pid=story_pid,
-                        process_image=story_image,
-                        preserve_dst_ip=bool(spec.hostname),
-                        preserve_explicit_payload=(
-                            spec.orig_bytes is not None
-                            or spec.resp_bytes is not None
-                            or tick_sequence_entry is not None
-                        ),
-                    )
-                attempt_count += 1
-
-            malicious_event["dst_ip"] = beacon_dst_ip
-            malicious_event["dst_port"] = spec.dst_port
-            malicious_event["interval"] = spec.interval
-            malicious_event["action"] = spec.action
-            term = spec.duration or spec.end_time or f"count={spec.count}"
-            malicious_event["termination"] = term
-            malicious_event["attempt_count"] = attempt_count
-            if getattr(spec, "ids_alerts", []):
-                malicious_event["ids_alerts"] = _ids_attachment_ground_truth(spec.ids_alerts)
-
-        elif spec.type == "dns_query":
-            # QTYPE name → numeric mapping
-            _QTYPE_MAP = {
-                "A": 1,
-                "AAAA": 28,
-                "TXT": 16,
-                "CNAME": 5,
-                "MX": 15,
-                "NULL": 10,
-                "SRV": 33,
-                "PTR": 12,
-            }
-            _RCODE_MAP = {"NOERROR": 0, "NXDOMAIN": 3, "SERVFAIL": 2, "REFUSED": 5}
-
-            from evidenceforge.events.contexts import DnsContext
-
-            qtype_num = _QTYPE_MAP.get(spec.qtype, 1)
-            rcode_num = _RCODE_MAP.get(spec.rcode, 0)
-
-            # Build answers list
-            answers = []
-            ttls = []
-            if spec.answer is not None:
-                answers = [spec.answer] if isinstance(spec.answer, str) else list(spec.answer)
-                ttl_val = float(spec.ttl) if spec.ttl is not None else float(rng.randint(60, 3600))
-                ttls = [ttl_val] * len(answers)
-            elif spec.rcode == "NOERROR" and spec.qtype in {"A", "AAAA"}:
-                resolver = getattr(self, "network_resolver", None)
-                if resolver is not None:
-                    resolved_query = resolver.resolve_host(spec.query, src_host=system.hostname)
-                    if resolved_query.ip:
-                        answers = [resolved_query.ip]
-                        ttl_val = (
-                            float(spec.ttl)
-                            if spec.ttl is not None
-                            else float(rng.randint(60, 3600))
-                        )
-                        ttls = [ttl_val]
-
-            # Resolve DNS server IP before choosing source-native DNS RTT so
-            # local resolvers do not get impossible multi-second timings.
-            query_src_ip = spec.source_ip or system.ip
-            dns_server_ips = activity_dns_resolver_ips(
-                self.activity_generator,
-                query_src_ip,
-            )
-            dns_server_ip = rng.choice(dns_server_ips)
-            from evidenceforge.generation.activity.generator import _dns_rtt
-
-            dns_ctx = DnsContext(
-                query=spec.query,
-                query_type=spec.qtype,
-                qtype=qtype_num,
-                rcode=spec.rcode,
-                rcode_num=rcode_num,
-                answers=answers,
-                TTLs=ttls,
-                preserve_ttls=spec.ttl is not None,
-                trans_id=rng.randint(1, 65535),
-                AA=False,
-                RD=True,
-                RA=True,
-                rejected=spec.rcode == "REFUSED",
-                rtt=_dns_rtt(rng, dns_server_ip),
-            )
-            authored_ids_alerts = _build_ids_alert_contexts(
-                getattr(spec, "ids_alerts", []),
-                time=time,
-                src_ip=query_src_ip,
-                dst_ip=dns_server_ip,
-                dst_port=53,
-                proto="udp",
-                rng=rng,
-                source="storyline_dns_query",
-            )
-
-            self.activity_generator.generate_connection(
-                src_ip=query_src_ip,
-                dst_ip=dns_server_ip,
-                time=time,
-                dst_port=53,
-                proto="udp",
-                service="dns",
-                dns=dns_ctx,
-                emit_dns=False,
-                orig_bytes=rng.randint(40, 100),
-                resp_bytes=rng.randint(80, 400) if spec.rcode == "NOERROR" else rng.randint(40, 80),
-                conn_state="SF",
-                duration=rng.uniform(0.001, 0.05),
-                ids_alerts=authored_ids_alerts,
-            )
-
-            malicious_event["query"] = spec.query
-            malicious_event["qtype"] = spec.qtype
-            malicious_event["rcode"] = spec.rcode
-            if getattr(spec, "ids_alerts", []):
-                malicious_event["ids_alerts"] = _ids_attachment_ground_truth(spec.ids_alerts)
-
-        elif spec.type == "web_scan":
-            malicious_event = WebScanActionBundle(
-                executor=self,
-                request=WebScanRequest(
-                    spec=spec,
-                    actor=actor,
-                    system=system,
-                    time=time,
-                    rng=rng,
-                    malicious_event=malicious_event,
-                ),
-            ).execute()
-
-        elif spec.type == "credential_spray":
-            # Timing
-            start = (
-                self._parse_storyline_time(spec.start_time) + authored_time_shift
-                if spec.start_time
-                else time
-            )
-            interval_sec = parse_duration(spec.interval).total_seconds()
-            duration_sec = None
-            count = spec.count
-            if spec.duration is not None:
-                duration_sec = parse_duration(spec.duration).total_seconds()
-            elif spec.end_time is not None:
-                end_dt = self._parse_storyline_time(spec.end_time) + authored_time_shift
-                duration_sec = (end_dt - start).total_seconds()
-
-            spray_src_ip = spec.source_ip or system.ip
-            accounts = spec.target_accounts
-            success_spec = spec.success
-            success_account = success_spec.get("account") if success_spec else None
-            success_after = success_spec.get("after", 0) if success_spec else 0
-            success_session_end_plan = (
-                self._authored_rdp_session_end_plan()
-                if success_account
-                and spec.logon_type == 10
-                and _get_os_category(system.os) == "windows"
-                and spray_src_ip not in {"-", system.ip}
-                else None
-            )
-
-            # Resolve target accounts — include service accounts as synthetic User
-            # objects so credential_spray targets resolve for both failed and success logons
-            from evidenceforge.models.scenario import User as _User
-
-            scenario_users = {u.username: u for u in self.scenario.environment.users}
-            ad_domain = self.scenario.environment.domain or "corp.local"
-            for svc_name in self.scenario.environment.service_accounts:
-                if svc_name not in scenario_users:
-                    scenario_users[svc_name] = _User(
-                        username=svc_name,
-                        full_name=svc_name,
-                        email=f"{svc_name}@{ad_domain}",
-                    )
-
-            # Only attach DC for Windows domain-account sprays — Linux SSH brute
-            # force or local-account attacks should not produce DC-side 4625/4776
-            dc_system = None
-            is_windows_target = "windows" in system.os.lower()
-            has_domain_account = any(acct in scenario_users for acct in accounts)
-            if is_windows_target and has_domain_account:
-                dcs = [
-                    s for s in self.scenario.environment.systems if s.type == "domain_controller"
-                ]
-                if dcs:
-                    # Deterministic DC per source IP (mimics AD DC Locator caching)
-                    dc_idx = _stable_seed(f"preferred_dc_{spray_src_ip}") % len(dcs)
-                    dc_system = dcs[dc_idx]
-
-            attempt_count = 0
-            for tick_time in _iter_periodic_ticks(
-                start,
-                interval_sec,
-                duration_sec,
-                count,
-                spec.jitter,
-                rng,
-                exclusive_end_time=getattr(self, "end_time", None),
-            ):
-                self.state_manager.set_current_time(tick_time)
-
-                # Success fires at exactly the requested attempt count,
-                # regardless of which account the pattern would have selected
-                if success_account and attempt_count == success_after:
-                    target_user = scenario_users.get(success_account, actor)
-                    self.activity_generator.generate_logon(
-                        user=target_user,
-                        system=system,
-                        time=tick_time,
-                        logon_type=spec.logon_type,
-                        source_ip=spray_src_ip,
-                        session_end_plan=success_session_end_plan,
-                    )
-                    attempt_count += 1
-                    malicious_event["success_account"] = success_account
-                    malicious_event["success_at_attempt"] = attempt_count
-                    break
-
-                # Select target account based on pattern
-                if spec.pattern == "spray":
-                    target_account = accounts[attempt_count % len(accounts)]
-                elif spec.pattern == "brute_force":
-                    target_account = accounts[
-                        min(
-                            attempt_count // max(1, (spec.count or 100) // len(accounts)),
-                            len(accounts) - 1,
-                        )
-                    ]
-                else:  # stuffing
-                    target_account = accounts[attempt_count % len(accounts)]
-
-                target_user = scenario_users.get(target_account, actor)
-
-                self.activity_generator.generate_failed_logon(
-                    user=target_user,
-                    system=system,
-                    time=tick_time,
-                    logon_type=spec.logon_type,
-                    source_ip=spray_src_ip,
-                    target_username=target_account,
-                    dc_system=dc_system,
-                )
-                attempt_count += 1
-
-            malicious_event["pattern"] = spec.pattern
-            malicious_event["target_accounts"] = accounts
-            malicious_event["attempt_count"] = attempt_count
-
-        elif spec.type == "dga_queries":
-            import random as _random
-
-            from evidenceforge.events.contexts import DnsContext
-
-            # Timing
-            start = self._parse_storyline_time(spec.start_time) if spec.start_time else time
-            interval_sec = parse_duration(spec.interval).total_seconds()
-            duration_sec = None
-            count = spec.count
-            if spec.duration is not None:
-                duration_sec = parse_duration(spec.duration).total_seconds()
-            elif spec.end_time is not None:
-                end_dt = self._parse_storyline_time(spec.end_time)
-                duration_sec = (end_dt - start).total_seconds()
-
-            # DGA RNG — separate from main rng for reproducibility
-            dga_seed = spec.seed if spec.seed is not None else rng.randint(0, 2**31)
-            dga_rng = _random.Random(dga_seed)
-
-            # Rcode distribution
-            rcode_dist = spec.rcode_distribution or {"NXDOMAIN": 0.95, "NOERROR": 0.05}
-            rcode_names = list(rcode_dist.keys())
-            rcode_weights = list(rcode_dist.values())
-
-            _RCODE_MAP = {"NOERROR": 0, "NXDOMAIN": 3, "SERVFAIL": 2, "REFUSED": 5}
-            _QTYPE_MAP = {"A": 1, "AAAA": 28, "TXT": 16, "CNAME": 5}
-
-            query_src_ip = spec.source_ip or system.ip
-            dns_server_ips = activity_dns_resolver_ips(
-                self.activity_generator,
-                query_src_ip,
-            )
-
-            query_count = 0
-            nxdomain_count = 0
-            domain_sample = []
-            for tick_time in _iter_periodic_ticks(
-                start,
-                interval_sec,
-                duration_sec,
-                count,
-                spec.jitter,
-                rng,
-                exclusive_end_time=getattr(self, "end_time", None),
-            ):
-                self.state_manager.set_current_time(tick_time)
-
-                # Generate random domain
-                label_len = dga_rng.randint(*spec.length_range)
-                label = "".join(dga_rng.choices(spec.charset, k=label_len))
-                domain = f"{label}{spec.tld}"
-
-                # Select rcode
-                rcode_name = dga_rng.choices(rcode_names, weights=rcode_weights, k=1)[0]
-                rcode_num = _RCODE_MAP.get(rcode_name, 3)
-
-                answers = []
-                ttls = []
-                if rcode_name == "NOERROR" and spec.answer_ip:
-                    answers = [spec.answer_ip]
-                    ttls = [float(dga_rng.randint(60, 3600))]
-                if rcode_name == "NXDOMAIN":
-                    nxdomain_count += 1
-
-                dns_server_ip = rng.choice(dns_server_ips)
-                from evidenceforge.generation.activity.generator import _dns_rtt
-
-                dns_ctx = DnsContext(
-                    query=domain,
-                    query_type="A",
-                    qtype=1,
-                    rcode=rcode_name,
-                    rcode_num=rcode_num,
-                    answers=answers,
-                    TTLs=ttls,
-                    trans_id=rng.randint(1, 65535),
-                    AA=False,
-                    RD=True,
-                    RA=True,
-                    rejected=False,
-                    rtt=_dns_rtt(rng, dns_server_ip),
-                )
-                authored_ids_alerts = _build_ids_alert_contexts(
-                    getattr(spec, "ids_alerts", []),
-                    time=tick_time,
-                    src_ip=query_src_ip,
-                    dst_ip=dns_server_ip,
-                    dst_port=53,
-                    proto="udp",
-                    rng=rng,
-                    source="storyline_dga_queries",
-                )
-
-                self.activity_generator.generate_connection(
-                    src_ip=query_src_ip,
-                    dst_ip=dns_server_ip,
-                    time=tick_time,
-                    dst_port=53,
-                    proto="udp",
-                    service="dns",
-                    dns=dns_ctx,
-                    emit_dns=False,
-                    orig_bytes=rng.randint(40, 100),
-                    resp_bytes=rng.randint(80, 400)
-                    if rcode_name == "NOERROR"
-                    else rng.randint(40, 80),
-                    conn_state="SF",
-                    duration=rng.uniform(0.001, 0.05),
-                    ids_alerts=authored_ids_alerts,
-                )
-                query_count += 1
-                if query_count % 500 == 0:
-                    self.state_manager.sweep_closed_connections()
-                if len(domain_sample) < 5:
-                    domain_sample.append(domain)
-
-            malicious_event["total_queries"] = query_count
-            malicious_event["nxdomain_count"] = nxdomain_count
-            malicious_event["domain_sample"] = domain_sample
-            malicious_event["tld"] = spec.tld
-            if getattr(spec, "ids_alerts", []):
-                malicious_event["ids_alerts"] = _ids_attachment_ground_truth(spec.ids_alerts)
-
-        elif spec.type == "dns_tunnel":
-            import base64 as _b64
-
-            from evidenceforge.events.contexts import DnsContext
-            from evidenceforge.generation.activity.network_params import (
-                dns_tunnel_rcode_weights,
-                dns_tunnel_response_templates,
-                dns_tunnel_rtt_range,
-                dns_tunnel_ttl_choices,
-            )
-
-            _QTYPE_MAP = {"TXT": 16, "NULL": 10, "CNAME": 5}
-            _RCODE_MAP = {"NOERROR": 0, "NXDOMAIN": 3, "SERVFAIL": 2, "REFUSED": 5}
-
-            # Timing
-            start = self._parse_storyline_time(spec.start_time) if spec.start_time else time
-            interval_sec = parse_duration(spec.interval).total_seconds()
-            duration_sec = None
-            count = spec.count
-            if spec.duration is not None:
-                duration_sec = parse_duration(spec.duration).total_seconds()
-            elif spec.end_time is not None:
-                end_dt = self._parse_storyline_time(spec.end_time)
-                duration_sec = (end_dt - start).total_seconds()
-
-            exclusive_end_time = getattr(self, "end_time", None)
-            min_rtt, max_rtt = dns_tunnel_rtt_range()
-            scenario_start_time = getattr(self, "start_time", None)
-            dns_window_start = (
-                max(ensure_utc(start), ensure_utc(scenario_start_time))
-                if scenario_start_time is not None
-                else ensure_utc(start)
-            )
-            dns_start_fence = _dns_periodic_exclusive_start_fence(
-                self.activity_generator,
-                window_start=dns_window_start,
-                exclusive_end_time=exclusive_end_time,
-                maximum_rtt_seconds=max_rtt,
-            )
-            tunnel_ticks = _iter_dns_tunnel_ticks(
-                start,
-                interval_sec,
-                duration_sec,
-                count,
-                spec.jitter,
-                rng,
-                exclusive_end_time=dns_start_fence,
-            )
-            # Freeze one real cadence tick before any background publication. This
-            # intentionally moves cadence RNG ahead of payload/background RNG so the
-            # owner cannot emit cover traffic for a campaign with no admitted query.
-            first_tick = next(tunnel_ticks, None)
-            if first_tick is None:
-                malicious_event["base_domain"] = spec.base_domain
-                malicious_event["encoding"] = spec.encoding
-                malicious_event["qtype"] = spec.qtype
-                malicious_event["total_queries"] = 0
-                malicious_event["bytes_exfiltrated"] = 0
-                if getattr(spec, "ids_alerts", []):
-                    malicious_event["ids_alerts"] = _ids_attachment_ground_truth(spec.ids_alerts)
-                return malicious_event
-
-            query_src_ip = spec.source_ip or system.ip
-            dns_server_ips = activity_dns_resolver_ips(
-                self.activity_generator,
-                query_src_ip,
-            )
-
-            # Generate or use payload
-            if spec.payload:
-                payload_bytes = spec.payload.encode("utf-8")
-            else:
-                payload_bytes = rng.randbytes(spec.payload_size)
-
-            # Calculate raw bytes that can fit in the visible label for each encoding.
-            if spec.encoding == "hex":
-                bytes_per_label = spec.label_length // 2
-            elif spec.encoding == "base32":
-                bytes_per_label = (spec.label_length * 5) // 8
-            else:  # base64
-                bytes_per_label = (spec.label_length * 3) // 4
-            bytes_per_label = max(1, bytes_per_label)
-
-            # Reserve visible label capacity for tunnel metadata before chunking the payload.
-            # Otherwise full-sized chunks would be encoded with metadata and truncated, causing
-            # GROUND_TRUTH.md to count bytes that never appeared in the emitted DNS label.
-            visible_nonce_len = 2
-            sequence_len = 4
-            payload_bytes_per_label = max(0, bytes_per_label - visible_nonce_len - sequence_len)
-
-            # Chunk only the bytes that can actually be emitted in the label. Very small labels
-            # still generate DNS traffic but carry no visible payload, so ground truth reports 0.
-            chunks: list[bytes]
-            if payload_bytes_per_label > 0:
-                chunks = [
-                    payload_bytes[i : i + payload_bytes_per_label]
-                    for i in range(0, len(payload_bytes), payload_bytes_per_label)
-                ]
-            else:
-                chunks = [b""]
-
-            qtype_num = _QTYPE_MAP.get(spec.qtype, 16)
-            response_templates = dns_tunnel_response_templates() or ["status={token}"]
-            response_primary_template = rng.choice(response_templates)
-            response_secondary_templates = [
-                template
-                for template in rng.sample(
-                    response_templates,
-                    k=min(len(response_templates), rng.randint(3, 6)),
-                )
-                if template != response_primary_template
-            ]
-            ttl_choices = dns_tunnel_ttl_choices()
-            campaign_ttl = _choose_dns_tunnel_campaign_ttl(ttl_choices, rng)
-            rcode_weights = dns_tunnel_rcode_weights()
-            rcode_names = list(rcode_weights)
-            rcode_values = [rcode_weights[name] for name in rcode_names]
-            total_bytes = 0
-            query_count = 0
-            chunk_idx = 0
-            tunnel_salt = rng.randbytes(4)
-
-            scenario = getattr(self, "scenario", None)
-            environment = getattr(scenario, "environment", None)
-            background_systems = [
-                candidate
-                for candidate in getattr(environment, "systems", [])
-                if getattr(candidate, "ip", "") and getattr(candidate, "ip", "") != query_src_ip
-            ]
-            background_window_sec = (
-                duration_sec
-                if duration_sec is not None
-                else interval_sec * float(count if count is not None else 120)
-            )
-            background_earliest = start - timedelta(seconds=240.0)
-            background_latest = start + timedelta(seconds=background_window_sec + 240.0)
-            if scenario_start_time is not None:
-                background_earliest = max(
-                    ensure_utc(background_earliest),
-                    ensure_utc(scenario_start_time),
-                )
-            if dns_start_fence is not None:
-                # Sample once from the valid intersection instead of retrying
-                # post-fence candidates and drifting the owner RNG.
-                background_latest = min(
-                    ensure_utc(background_latest),
-                    dns_start_fence - timedelta(microseconds=1),
-                )
-            background_span_sec = max(
-                0.0,
-                (ensure_utc(background_latest) - ensure_utc(background_earliest)).total_seconds(),
-            )
-            if (
-                background_systems
-                and background_window_sec > 0
-                and ensure_utc(background_earliest) <= ensure_utc(background_latest)
-            ):
-                background_count = min(36, max(12, len(background_systems) * 2 + rng.randint(3, 9)))
-                for _ in range(background_count):
-                    bg_system = rng.choice(background_systems)
-                    bg_query, bg_answer, bg_ttl = _dns_tunnel_background_txt_record(rng)
-                    bg_rtt = rng.uniform(min_rtt, max_rtt)
-                    bg_dns = DnsContext(
-                        query=bg_query,
-                        query_type="TXT",
-                        qtype=16,
-                        rcode="NOERROR",
-                        rcode_num=0,
-                        answers=[bg_answer],
-                        TTLs=[float(bg_ttl)],
-                        trans_id=rng.randint(1, 65535),
-                        AA=False,
-                        RD=True,
-                        RA=True,
-                        rejected=False,
-                        rtt=bg_rtt,
-                    )
-                    bg_time = background_earliest + timedelta(
-                        seconds=rng.uniform(0.0, background_span_sec)
-                    )
-                    self.activity_generator.generate_connection(
-                        src_ip=bg_system.ip,
-                        dst_ip=rng.choice(dns_server_ips),
-                        time=bg_time,
-                        dst_port=53,
-                        proto="udp",
-                        service="dns",
-                        dns=bg_dns,
-                        emit_dns=False,
-                        resp_bytes=max(90, len(bg_query) + len(bg_answer) + rng.randint(35, 120)),
-                        duration=bg_rtt,
-                        source_system=bg_system,
-                    )
-
-            for tick_time in itertools.chain((first_tick,), tunnel_ticks):
-                self.state_manager.set_current_time(tick_time)
-
-                if spec.label_length >= 24:
-                    min_label_length = max(14, int(spec.label_length * 0.45))
-                    label_length = int(
-                        rng.triangular(min_label_length, spec.label_length, spec.label_length - 4)
-                    )
-                elif spec.label_length >= 20:
-                    label_length = rng.randint(max(16, spec.label_length - 8), spec.label_length)
-                else:
-                    label_length = spec.label_length
-                if spec.encoding == "hex":
-                    effective_bytes_per_label = label_length // 2
-                elif spec.encoding == "base32":
-                    effective_bytes_per_label = (label_length * 5) // 8
-                else:  # base64
-                    effective_bytes_per_label = (label_length * 3) // 4
-                effective_bytes_per_label = max(1, effective_bytes_per_label)
-
-                chunk = chunks[chunk_idx % len(chunks)]
-                chunk_idx += 1
-                sequence_mask = random.Random(
-                    _stable_seed(
-                        f"dns_tunnel_seq:{spec.base_domain}:{tunnel_salt.hex()}:{query_count}"
-                    )
-                ).getrandbits(32)
-                sequence = (query_count ^ sequence_mask).to_bytes(4, "big", signed=False)
-                visible_nonce = rng.randbytes(visible_nonce_len)
-                effective_payload_capacity = max(
-                    0,
-                    effective_bytes_per_label - visible_nonce_len - sequence_len,
-                )
-                visible_payload = chunk[: min(payload_bytes_per_label, effective_payload_capacity)]
-                pad_len = max(
-                    0,
-                    effective_bytes_per_label
-                    - len(visible_nonce)
-                    - len(visible_payload)
-                    - len(sequence),
-                )
-                padded_chunk = visible_nonce + visible_payload + rng.randbytes(pad_len) + sequence
-
-                # Encode chunk
-                if spec.encoding == "hex":
-                    encoded = padded_chunk.hex()
-                elif spec.encoding == "base32":
-                    encoded = _b64.b32encode(padded_chunk).decode("ascii").rstrip("=").lower()
-                else:  # base64
-                    encoded = (
-                        _b64.urlsafe_b64encode(padded_chunk).decode("ascii").rstrip("=").lower()
-                    )
-
-                # Truncate to label_length
-                encoded = encoded[:label_length]
-                query_labels = [
-                    encoded,
-                    *_dns_tunnel_extra_labels(
-                        query_count,
-                        random.Random(
-                            _stable_seed(
-                                "dns_tunnel_extra_labels:"
-                                f"{spec.base_domain}:{tunnel_salt.hex()}:{query_count}"
-                            )
-                        ),
-                    ),
-                ]
-                tunnel_query = ".".join([*query_labels, spec.base_domain])
-
-                rcode_name = rng.choices(rcode_names, weights=rcode_values, k=1)[0]
-                rcode_num = _RCODE_MAP.get(rcode_name, 0)
-                answers: list[str] = []
-                ttls: list[float] = []
-                if rcode_name == "NOERROR":
-                    # TXT responses carry data back; CNAME/NULL are smaller.
-                    if spec.qtype == "TXT":
-                        resp_bytes = rng.randint(140, 2400)
-                    else:
-                        resp_bytes = rng.randint(50, 240)
-                    token_rng = random.Random(
-                        _stable_seed(
-                            f"dns_tunnel_response:{spec.base_domain}:{query_count}:"
-                            f"{tunnel_salt.hex()}"
-                        )
-                    )
-                    token_bytes = token_rng.randbytes(token_rng.randint(3, 10))
-                    token_style = token_rng.choice(["hex", "base32", "base64url"])
-                    if token_style == "base32":
-                        response_token = _b64.b32encode(token_bytes).decode("ascii").rstrip("=")
-                    elif token_style == "base64url":
-                        response_token = (
-                            _b64.urlsafe_b64encode(token_bytes).decode("ascii").rstrip("=")
-                        )
-                    else:
-                        response_token = token_bytes.hex()
-                    response_ttl = _choose_dns_tunnel_response_ttl(
-                        ttl_choices,
-                        campaign_ttl,
-                        rng,
-                    )
-                    response_template = _choose_dns_tunnel_response_template(
-                        response_templates,
-                        response_primary_template,
-                        response_secondary_templates,
-                        rng,
-                    )
-                    answers = [
-                        _render_dns_tunnel_response_template(
-                            response_template,
-                            token=response_token,
-                            query_count=query_count,
-                            ttl=response_ttl,
-                            rng=rng,
-                        )
-                    ]
-                    ttls = [response_ttl]
-                else:
-                    resp_bytes = rng.randint(55, 180)
-
-                dns_ctx = DnsContext(
-                    query=tunnel_query,
-                    query_type=spec.qtype,
-                    qtype=qtype_num,
-                    rcode=rcode_name,
-                    rcode_num=rcode_num,
-                    answers=answers,
-                    TTLs=ttls,
-                    trans_id=rng.randint(1, 65535),
-                    AA=False,
-                    RD=True,
-                    RA=True,
-                    rejected=False,
-                    rtt=rng.uniform(min_rtt, max_rtt),
-                )
-
-                dns_server_ip = rng.choice(dns_server_ips)
-                authored_ids_alerts = _build_ids_alert_contexts(
-                    getattr(spec, "ids_alerts", []),
-                    time=tick_time,
-                    src_ip=query_src_ip,
-                    dst_ip=dns_server_ip,
-                    dst_port=53,
-                    proto="udp",
-                    rng=rng,
-                    source="storyline_dns_tunnel",
-                )
-                self.activity_generator.generate_connection(
-                    src_ip=query_src_ip,
-                    dst_ip=dns_server_ip,
-                    time=tick_time,
-                    dst_port=53,
-                    proto="udp",
-                    service="dns",
-                    dns=dns_ctx,
-                    emit_dns=False,
-                    resp_bytes=resp_bytes,
-                    duration=dns_ctx.rtt,
-                    ids_alerts=authored_ids_alerts,
-                )
-                total_bytes += len(visible_payload)
-                query_count += 1
-
-            malicious_event["base_domain"] = spec.base_domain
-            malicious_event["encoding"] = spec.encoding
-            malicious_event["qtype"] = spec.qtype
-            malicious_event["total_queries"] = query_count
-            malicious_event["bytes_exfiltrated"] = total_bytes
-            if getattr(spec, "ids_alerts", []):
-                malicious_event["ids_alerts"] = _ids_attachment_ground_truth(spec.ids_alerts)
-
-        elif spec.type == "explicit_credentials":
-            story_pid, _story_image = self._last_storyline_process_for_system(system)
-            self.activity_generator.generate_explicit_credentials(
-                user=actor,
-                system=system,
-                time=time,
-                target_username=spec.target_username,
-                target_server=spec.target_server or system.hostname,
-                process_name=spec.process_name or r"C:\Windows\System32\runas.exe",
-                process_pid=story_pid if story_pid > 0 else 0,
-                source_ip=spec.source_ip or "",
-            )
-            malicious_event["target_username"] = spec.target_username
-            malicious_event["target_server"] = spec.target_server
-
-        elif spec.type == "workstation_lock":
-            sessions = self.state_manager.get_sessions_for_user(actor.username)
-            session = max(
-                (
-                    s
-                    for s in sessions
-                    if s.system == system.hostname
-                    and s.logon_type in (2, 11)
-                    and s.session_kind not in {"network", "service", "rdp", "ssh"}
-                    and s.start_time <= time
-                ),
-                key=lambda s: s.start_time,
-                default=None,
-            )
-            logon_id = session.logon_id if session else "0x0"
-            result = self.activity_generator.generate_workstation_lock(
-                user=actor,
-                system=system,
-                time=time,
-                logon_id=logon_id,
-            )
-            if not result.emitted:
-                malicious_event["skipped_reason"] = (
-                    result.skipped_reason or "workstation_lock_not_emitted"
-                )
-
-        elif spec.type == "workstation_unlock":
-            sessions = self.state_manager.get_sessions_for_user(actor.username)
-            session = max(
-                (
-                    s
-                    for s in sessions
-                    if s.system == system.hostname
-                    and s.logon_type in (2, 11)
-                    and s.session_kind not in {"network", "service", "rdp", "ssh"}
-                    and s.start_time <= time
-                ),
-                key=lambda s: s.start_time,
-                default=None,
-            )
-            logon_id = session.logon_id if session else "0x0"
-            self.activity_generator.generate_workstation_unlock(
-                user=actor,
-                system=system,
-                time=time,
-                logon_id=logon_id,
-            )
-
-        elif spec.type == "spillage":
-            from evidenceforge.generation.spillage import HTTP_SURFACES
-
-            cluster_id = malicious_event["storyline_cluster_id"]
-            # Monotonic per-generation sequence makes every spillage value unique
-            # (deterministic order); eval reads the emitted value from ground truth.
-            seq = getattr(self, "_spillage_seq", 0)
-            self._spillage_seq = seq + 1
-            spill_logon_id = None
-            spill_target = None
-            spill_scheme = spec.scheme
-            # process_command_line spills run as a standalone process-execution
-            # record with a durable unique PID, using local carriers only (no
-            # implied outbound network). We deliberately do NOT attach the spill to
-            # an interactive shell session: a foreground bash child competes for
-            # that shell's serialized command timeline and, under heavy baseline,
-            # can be shifted and dropped during eCAR post-flush normalization —
-            # leaving a labeled-but-unwritten credential (a phantom). A standalone
-            # process keeps a stable, unique identity and always lands.
-            if spec.surface == "process_command_line":
-                spill_logon_id = self._resolve_storyline_process_spill_logon_id(
-                    actor,
-                    system,
-                    time,
-                    rng,
-                )
-            if spec.surface in HTTP_SURFACES:
-                # The credential leaks into a web server's access log; pick the
-                # destination web server (the validator requires one to exist).
-                spill_target, spill_scheme = self._select_web_server_for_spillage(
-                    system, spec.scheme
-                )
-            info = self.activity_generator.generate_spillage(
-                user=actor,
-                system=system,
-                time=time,
-                surface=spec.surface,
-                family=spec.family,
-                value=spec.value,
-                scheme=spill_scheme,
-                seed_key=f"spillage:{cluster_id}:{seq}:{spec.surface}:{spec.family or 'literal'}",
-                logon_id=spill_logon_id,
-                target_system=spill_target,
-            )
-            malicious_event["surface"] = info["surface"]
-            malicious_event["family"] = info["family"]
-            if info.get("skipped_reason"):
-                # Credential was not emitted (e.g. dwell-shifted past the window);
-                # mark it so no phantom ground-truth record is written.
-                malicious_event["skipped_reason"] = info["skipped_reason"]
-            else:
-                malicious_event["value"] = info["value"]
-                malicious_event["rendered_value"] = info["rendered_value"]
-                malicious_event["expected_sources"] = info["expected_sources"]
-                # Reflect the actual emitted time (bash dwell scheduling may shift it).
-                malicious_event["time"] = info["time"]
-                if info.get("target_system"):
-                    # http_* surfaces land on the destination web server's access
-                    # log; record its FQDN (as the generator names the access-log
-                    # directory) so ground truth and eval can locate the trace.
-                    malicious_event["target_system"] = info["target_system"]
-                if info.get("scheme"):
-                    malicious_event["scheme"] = info["scheme"]
-
-        elif spec.type == "adversarial_payload":
-            from evidenceforge.generation.adversarial_payload import HTTP_SURFACES
-
-            cluster_id = malicious_event["storyline_cluster_id"]
-            # Monotonic per-generation sequence makes every synthesized payload
-            # unique (deterministic order); eval reads the emitted value from
-            # ground truth.
-            seq = getattr(self, "_adversarial_payload_seq", 0)
-            self._adversarial_payload_seq = seq + 1
-            payload_logon_id = None
-            payload_target = None
-            payload_scheme = None
-            # A process_command_line payload runs as a standalone process-execution
-            # record; resolve canonical session ownership (same as a spillage process
-            # spill) so it gets a non-shell parent and a stable, unique identity.
-            if spec.surface == "process_command_line":
-                payload_logon_id = self._resolve_storyline_process_spill_logon_id(
-                    actor,
-                    system,
-                    time,
-                    rng,
-                )
-            if spec.surface in HTTP_SURFACES:
-                # The payload rides to a web server's access log; pick the destination.
-                # An authored `scheme:` (http/https) forces the transport; otherwise the
-                # server's supported scheme decides (https preferred, else http).
-                payload_target, payload_scheme = self._select_web_server_for_spillage(
-                    system, spec.scheme
-                )
-            info = self.activity_generator.generate_adversarial_payload(
-                user=actor,
-                system=system,
-                time=time,
-                surface=spec.surface,
-                family=spec.family,
-                value=spec.value,
-                scheme=payload_scheme,
-                seed_key=f"adversarial_payload:{cluster_id}:{seq}:{spec.surface}:{spec.family or 'literal'}",
-                logon_id=payload_logon_id,
-                target_system=payload_target,
-            )
-            malicious_event["surface"] = info["surface"]
-            malicious_event["family"] = info["family"]
-            if info.get("skipped_reason"):
-                malicious_event["skipped_reason"] = info["skipped_reason"]
-            else:
-                malicious_event["value"] = info["value"]
-                malicious_event["rendered_value"] = info["rendered_value"]
-                malicious_event["expected_sources"] = info["expected_sources"]
-                malicious_event["encoding"] = info["encoding"]
-                malicious_event["time"] = info["time"]
-                if info.get("target_system"):
-                    malicious_event["target_system"] = info["target_system"]
-                if info.get("scheme"):
-                    malicious_event["scheme"] = info["scheme"]
-                if info.get("callback_host"):
-                    malicious_event["callback_host"] = info["callback_host"]
-                if info.get("weakness_class"):
-                    malicious_event["weakness_class"] = info["weakness_class"]
-                if info.get("expected_defender_signal"):
-                    malicious_event["expected_defender_signal"] = info["expected_defender_signal"]
-                if info.get("ids_alert"):
-                    malicious_event["ids_alert"] = info["ids_alert"]
-                # Pivot anchors to the exact evidence row (dst tuple for http, pid for
-                # process) so an analyst can jump from the payload record to the source.
-                for pivot_key in ("dst_ip", "dst_port", "pid"):
-                    if info.get(pivot_key) is not None:
-                        malicious_event[pivot_key] = info[pivot_key]
-
-        elif spec.type == "raw":
-            self.activity_generator.generate_raw(
-                time=time,
-                target_format=spec.target_format,
-                fields=spec.fields,
-                system=system,
-            )
-            malicious_event["target_format"] = spec.target_format
-
-        return malicious_event
+                dispatcher=dispatcher,
+                malicious_event=malicious_event,
+                _ground_truth_uid=_ground_truth_uid,
+            ),
+        )
 
     def _execute_port_scan_bundle(self, request: PortScanRequest) -> dict[str, Any]:
         """Expand a port-scan action bundle through the existing storyline adapter."""
